update: refactor utility code files
