+++ app/src/main/java/com/yukisoffd/lyracode/ai/agent/OpenAiAgent.kt
@@ -13,26 +13,16 @@ import com.yukisoffd.lyracode.R
 import com.yukisoffd.lyracode.data.ApiProfile
 import com.yukisoffd.lyracode.data.AppSettings
 import com.yukisoffd.lyracode.data.BackupManager
-import com.yukisoffd.lyracode.data.BackupOptions
 import com.yukisoffd.lyracode.data.ChatMessage
 import com.yukisoffd.lyracode.data.ConversationStore
 import com.yukisoffd.lyracode.data.DeepSeekV3Tokenizer
-import com.yukisoffd.lyracode.data.EmailServerConfig
-import com.yukisoffd.lyracode.data.FileTransferServerConfig
 import com.yukisoffd.lyracode.data.McpServerConfig
 import com.yukisoffd.lyracode.data.McpToolDefinition
-import com.yukisoffd.lyracode.data.MemoryEntry
 import com.yukisoffd.lyracode.data.MediaGenerationKind
-import com.yukisoffd.lyracode.data.MiniServerConfig
-import com.yukisoffd.lyracode.data.SkillPack
 import com.yukisoffd.lyracode.data.SubAgentConfig
-import com.yukisoffd.lyracode.data.SshServerConfig
-import com.yukisoffd.lyracode.data.WebDavServerConfig
 import com.yukisoffd.lyracode.debian.ProotCommandExecutor
 import com.yukisoffd.lyracode.filetransfer.FileTransferClient
 import com.yukisoffd.lyracode.email.EmailClient
-import com.yukisoffd.lyracode.email.EmailComposeRequest
-import com.yukisoffd.lyracode.email.OutgoingAttachment
 import com.yukisoffd.lyracode.mcp.McpClientManager
 import com.yukisoffd.lyracode.server.MiniServerManager
 import com.yukisoffd.lyracode.ssh.SshExecutor
@@ -40,15 +30,12 @@ import com.yukisoffd.lyracode.system.InstalledAppCollector
 import com.yukisoffd.lyracode.system.SystemCommandExecutor
 import com.yukisoffd.lyracode.tasks.DownloadTaskManager
 import com.yukisoffd.lyracode.tasks.DownloadTaskRequest
-import com.yukisoffd.lyracode.tasks.ScheduledTask
 import com.yukisoffd.lyracode.tasks.ScheduledTaskManager
-import com.yukisoffd.lyracode.tasks.ScheduledTaskType
 import com.yukisoffd.lyracode.termux.TermuxExecutor
 import com.yukisoffd.lyracode.uiText
 import com.yukisoffd.lyracode.webdav.WebDavClient
 import com.yukisoffd.lyracode.workspace.GlobalFileManager
 import com.yukisoffd.lyracode.workspace.NativeFileManager
-import com.yukisoffd.lyracode.workspace.WorkspaceFile
 import com.yukisoffd.lyracode.workspace.WorkspaceManager
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Dispatchers
@@ -68,243 +55,12 @@ import java.io.IOException
 import java.net.URI
 import java.security.MessageDigest
 import java.text.SimpleDateFormat
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
-import java.time.format.DateTimeParseException
 import java.util.Date
 import java.util.Locale
 import java.util.concurrent.ConcurrentHashMap
 import java.util.TimeZone
 import java.util.concurrent.TimeUnit
 
-
-internal val HISTORY_COMPRESSION_SCHEMA_V2 = """
-    LYRA_STRUCTURED_CONTEXT_V2
-    current_goal:
-    - ...
-    confirmed_facts:
-    - ...
-    constraints_and_preferences:
-    - ...
-    decisions_and_rationale:
-    - ...
-    completed_tasks:
-    - ...
-    pending_tasks:
-    - ...
-    important_artifacts:
-    - files, paths, code symbols, commands, IDs, URLs, configuration values, and outputs
-    errors_and_attempts:
-    - error, attempted remedy, and result
-    attention_items:
-    - risks, caveats, assumptions, conflicts, and details that must not be lost
-    next_actions:
-    - ...
-    open_questions:
-    - ...
-""".trimIndent()
-
-internal fun splitCompressionTranscript(transcript: String, requestedChunkCount: Int): List<String> {
-    if (transcript.isEmpty()) return emptyList()
-    val codePointCount = transcript.codePointCount(0, transcript.length)
-    val chunkCount = requestedChunkCount.coerceAtLeast(1).coerceAtMost(codePointCount)
-    if (chunkCount == 1) return listOf(transcript)
-    val chunks = ArrayList<String>(chunkCount)
-    var start = 0
-    repeat(chunkCount) { index ->
-        val chunksLeft = chunkCount - index
-        if (chunksLeft == 1) {
-            chunks += transcript.substring(start)
-            return@repeat
-        }
-        val remainingLength = transcript.length - start
-        val idealEnd = start + (remainingLength + chunksLeft - 1) / chunksLeft
-        val maxEnd = transcript.length - (chunksLeft - 1)
-        val searchRadius = minOf(384, maxOf(24, (idealEnd - start) / 8))
-        val forwardEnd = transcript.indexOf('\n', idealEnd)
-            .takeIf { it >= 0 && it + 1 <= maxEnd && it - idealEnd <= searchRadius }
-            ?.plus(1)
-        val backwardEnd = transcript.lastIndexOf('\n', idealEnd - 1)
-            .takeIf { it >= start && idealEnd - (it + 1) <= searchRadius }
-            ?.plus(1)
-        var end = listOfNotNull(forwardEnd, backwardEnd)
-            .minByOrNull { kotlin.math.abs(it - idealEnd) }
-            ?: idealEnd
-        if (end < transcript.length && end > start &&
-            Character.isHighSurrogate(transcript[end - 1]) && Character.isLowSurrogate(transcript[end])
-        ) {
-            end = if (end + 1 <= maxEnd) end + 1 else end - 1
-        }
-        end = end.coerceIn(start + 1, maxEnd)
-        chunks += transcript.substring(start, end)
-        start = end
-    }
-    return chunks
-}
-
-internal fun extractModelResponseText(root: JSONObject, apiFormat: String, useResponsesApi: Boolean = false): String {
-    val primary = when (apiFormat) {
-        ApiProfile.API_FORMAT_ANTHROPIC -> extractAnthropicResponseText(root)
-        ApiProfile.API_FORMAT_GEMINI -> extractGeminiResponseText(root)
-        else -> if (useResponsesApi) extractResponsesApiText(root) else extractOpenAiChatText(root)
-    }
-    if (primary.isNotBlank()) return primary
-    val fallbacks = when (apiFormat) {
-        ApiProfile.API_FORMAT_ANTHROPIC -> listOf(extractOpenAiChatText(root), extractResponsesApiText(root))
-        ApiProfile.API_FORMAT_GEMINI -> listOf(extractOpenAiChatText(root), extractResponsesApiText(root))
-        else -> if (useResponsesApi) listOf(extractOpenAiChatText(root)) else listOf(extractResponsesApiText(root))
-    }
-    fallbacks.firstOrNull { it.isNotBlank() }?.let { return it }
-    root.optJSONObject("response")?.let { wrapped ->
-        extractModelResponseText(wrapped, apiFormat, useResponsesApi).takeIf { it.isNotBlank() }?.let { return it }
-    }
-    root.optJSONObject("data")?.let { wrapped ->
-        extractModelResponseText(wrapped, apiFormat, useResponsesApi).takeIf { it.isNotBlank() }?.let { return it }
-    }
-    return ""
-}
-
-private fun extractAnthropicResponseText(root: JSONObject): String {
-    return extractVisibleText(root.opt("content")).ifBlank { root.optString("completion") }
-}
-
-private fun extractGeminiResponseText(root: JSONObject): String {
-    val candidates = root.optJSONArray("candidates") ?: return ""
-    return buildList {
-        for (index in 0 until candidates.length()) {
-            val candidate = candidates.optJSONObject(index) ?: continue
-            val content = candidate.optJSONObject("content")
-            extractVisibleText(content?.opt("parts")).takeIf { it.isNotBlank() }?.let(::add)
-            candidate.optString("output").takeIf { it.isNotBlank() }?.let(::add)
-            candidate.optString("text").takeIf { it.isNotBlank() }?.let(::add)
-        }
-    }.distinct().joinToString("\n")
-}
-
-private fun extractOpenAiChatText(root: JSONObject): String {
-    val choices = root.optJSONArray("choices") ?: return ""
-    return buildList {
-        for (index in 0 until choices.length()) {
-            val choice = choices.optJSONObject(index) ?: continue
-            val message = choice.optJSONObject("message")
-            extractVisibleText(message?.opt("content")).takeIf { it.isNotBlank() }?.let(::add)
-            choice.optString("text").takeIf { it.isNotBlank() }?.let(::add)
-        }
-    }.distinct().joinToString("\n")
-}
-
-private fun extractResponsesApiText(root: JSONObject): String {
-    root.optString("output_text").takeIf { it.isNotBlank() }?.let { return it }
-    val output = root.optJSONArray("output") ?: return ""
-    return buildList {
-        for (index in 0 until output.length()) {
-            val item = output.optJSONObject(index) ?: continue
-            if (item.optString("type") == "reasoning") continue
-            extractVisibleText(item.opt("content")).takeIf { it.isNotBlank() }?.let(::add)
-            extractVisibleText(item.opt("text")).takeIf { it.isNotBlank() }?.let(::add)
-        }
-    }.distinct().joinToString("\n")
-}
-
-private fun extractVisibleText(value: Any?): String = when (value) {
-    null, JSONObject.NULL -> ""
-    is String -> value.takeUnless { it.equals("null", ignoreCase = true) }.orEmpty()
-    is JSONArray -> buildList {
-        for (index in 0 until value.length()) {
-            extractVisibleText(value.opt(index)).takeIf { it.isNotBlank() }?.let(::add)
-        }
-    }.joinToString("\n")
-    is JSONObject -> {
-        when (value.optString("type")) {
-            "reasoning", "thinking", "function_call", "tool_call" -> ""
-            else -> listOf("text", "output_text", "content", "value")
-                .asSequence()
-                .map { extractVisibleText(value.opt(it)) }
-                .firstOrNull { it.isNotBlank() }
-                .orEmpty()
-        }
-    }
-    else -> ""
-}
-
-
-private data class ToolCall(
-    val id: String,
-    val name: String,
-    val arguments: JSONObject,
-    val rawArguments: String,
-) {
-    fun toJson(): JSONObject {
-        return JSONObject()
-            .put("id", id)
-            .put("type", "function")
-            .put(
-                "function",
-                JSONObject()
-                    .put("name", name)
-                    .put("arguments", rawArguments),
-            )
-    }
-}
-
-private data class StreamingResult(
-    val content: String,
-    val thinking: String,
-    val rawMessage: JSONObject,
-    val toolCalls: List<ToolCall>,
-    val tokensPerSecond: Double = 0.0,
-    val deepSeekCacheHitRate: Double? = null,
-    val fromCache: Boolean = false,
-)
-
-private class ToolCallBuilder {
-    var id: String = ""
-    var name: String = ""
-    val arguments = StringBuilder()
-
-    fun toToolCall(index: Int): ToolCall? {
-        if (name.isBlank()) return null
-        val raw = arguments.toString().ifBlank { "{}" }
-        val parsed = runCatching { JSONObject(raw) }.getOrElse { JSONObject() }
-        return ToolCall(id.ifBlank { "tool_$index" }, name, parsed, raw)
-    }
-
-    fun toJson(): JSONObject {
-        return JSONObject()
-            .put("id", id)
-            .put("type", "function")
-            .put(
-                "function",
-                JSONObject()
-                    .put("name", name)
-                    .put("arguments", arguments.toString()),
-            )
-    }
-}
-
-private class AnthropicBlockBuilder {
-    var type: String = ""
-    var id: String = ""
-    var name: String = ""
-    val text = StringBuilder()
-    val thinking = StringBuilder()
-    val input = StringBuilder()
-}
-
-private data class ToolExecution(
-    val content: String,
-    val fileChanges: List<FileDiff> = emptyList(),
-    val ok: Boolean = true,
-)
-
-private data class SubAgentExecutionContext(
-    val owner: SubAgentWriteOwner,
-    val agent: SubAgentConfig,
-    val readOnly: Boolean,
-    val writePaths: Set<String>,
-)
 
 class OpenAiAgent(
     private val context: Context,
@@ -370,6 +126,40 @@ class OpenAiAgent(
     private val subAgentContexts = ConcurrentHashMap<Long, SubAgentExecutionContext>()
     private val subAgentWriteCoordinator = SubAgentWriteCoordinator()
     private val emailClient = EmailClient(context)
+    private val fileTools = AgentFileToolHandler(
+        nativeFileManager = nativeFileManager,
+        globalFileManager = globalFileManager,
+        onFileEdit = { fileEditHandler(it) },
+        onFileMutation = { fileMutationHandler(it) },
+        onFileActivity = { fileActivityHandler(it) },
+    )
+    private val knowledgeTools = AgentKnowledgeToolHandler(settings, conversationStore)
+    private val automationTools = AgentAutomationToolHandler(
+        settings = settings,
+        scheduledTaskManager = scheduledTaskManager,
+        miniServerManager = miniServerManager,
+        parseTime = knowledgeTools::parseAgentTime,
+    )
+    private val configTools = AgentConfigToolHandler(
+        settings = settings,
+        mcpClientManager = mcpClientManager,
+        webDavClient = webDavClient,
+        fileTransferClient = fileTransferClient,
+        client = client,
+        configurableAgentTools = CONFIGURABLE_AGENT_TOOLS,
+        onConfigChanged = { configChangedHandler() },
+    )
+    private val remoteTools = AgentRemoteToolHandler(
+        settings = settings,
+        nativeFileManager = nativeFileManager,
+        globalFileManager = globalFileManager,
+        emailClient = emailClient,
+        sshExecutor = sshExecutor,
+        webDavClient = webDavClient,
+        fileTransferClient = fileTransferClient,
+        backupManager = backupManager,
+        onConfigChanged = { configChangedHandler() },
+    )
 
     suspend fun chat(
         conversationId: Long,
@@ -2231,21 +2021,21 @@ class OpenAiAgent(
                 when (call.name) {
                 "list_directory" -> nativeFileManager.listDirectory(args.optString("path"))
                     .fold({ ToolExecution(it.toAgentText()) }, { throw it })
-                "read_file" -> readFileWithActivity(args.getString("path"), globalStorage = false)
-                "read_file_lines" -> ToolExecution(readFileLines(args, globalStorage = false))
-                "write_file" -> writeFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
-                "edit_file" -> editFileWithDiff(args, globalStorage = false)
-                "append_file" -> appendFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
+                "read_file" -> fileTools.readFileWithActivity(args.getString("path"), globalStorage = false)
+                "read_file_lines" -> ToolExecution(fileTools.readFileLines(args, globalStorage = false))
+                "write_file" -> fileTools.writeFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
+                "edit_file" -> fileTools.editFileWithDiff(args, globalStorage = false)
+                "append_file" -> fileTools.appendFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
                 "create_folder" -> ToolExecution(nativeFileManager.createFolder(args.getString("path")).getOrThrow())
-                "delete_file_or_folder" -> deleteWithDiff(args.getString("path"))
-                "rename_move" -> renameMoveWithDiff(args.getString("from"), args.getString("to"))
+                "delete_file_or_folder" -> fileTools.deleteWithDiff(args.getString("path"))
+                "rename_move" -> fileTools.renameMoveWithDiff(args.getString("from"), args.getString("to"))
                 "global_list_directory" -> globalFileManager.listDirectory(args.optString("path"))
                     .fold({ ToolExecution(it.toAgentText()) }, { throw it })
-                "global_read_file" -> readFileWithActivity(args.getString("path"), globalStorage = true)
-                "global_read_file_lines" -> ToolExecution(readFileLines(args, globalStorage = true))
-                "global_write_file" -> globalWriteFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
-                "global_edit_file" -> editFileWithDiff(args, globalStorage = true)
-                "global_append_file" -> globalAppendFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
+                "global_read_file" -> fileTools.readFileWithActivity(args.getString("path"), globalStorage = true)
+                "global_read_file_lines" -> ToolExecution(fileTools.readFileLines(args, globalStorage = true))
+                "global_write_file" -> fileTools.globalWriteFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
+                "global_edit_file" -> fileTools.editFileWithDiff(args, globalStorage = true)
+                "global_append_file" -> fileTools.globalAppendFileWithDiff(args.getString("path"), args.toolTextArgument("content"))
                 "global_create_folder" -> ToolExecution(globalFileManager.createFolder(args.getString("path")).getOrThrow())
                 "global_delete_file_or_folder" -> ToolExecution(globalFileManager.delete(args.getString("path")).getOrThrow())
                 "global_rename_move" -> ToolExecution(globalFileManager.renameMove(args.getString("from"), args.getString("to")).getOrThrow())
@@ -2254,20 +2044,20 @@ class OpenAiAgent(
                 OCR_TOOL_NAME -> extractImageText(conversationId, args)
                 "generate_image", "generate_video", "generate_music", "generate_audio" ->
                     generateMedia(conversationId, mediaGenerationKindForTool(call.name)!!, args)
-                "manage_scheduled_tasks" -> ToolExecution(manageScheduledTasks(args))
-                "search_conversation_history" -> ToolExecution(searchConversationHistory(args))
-                "read_conversation_history" -> ToolExecution(readConversationHistory(args))
-                "read_memories" -> ToolExecution(readMemories(args))
-                "save_memory" -> ToolExecution(saveMemory(args))
-                "update_memory" -> ToolExecution(updateMemory(args))
-                "delete_memory" -> ToolExecution(deleteMemory(args))
+                "manage_scheduled_tasks" -> ToolExecution(automationTools.manageScheduledTasks(args))
+                "search_conversation_history" -> ToolExecution(knowledgeTools.searchConversationHistory(args))
+                "read_conversation_history" -> ToolExecution(knowledgeTools.readConversationHistory(args))
+                "read_memories" -> ToolExecution(knowledgeTools.readMemories(args))
+                "save_memory" -> ToolExecution(knowledgeTools.saveMemory(args))
+                "update_memory" -> ToolExecution(knowledgeTools.updateMemory(args))
+                "delete_memory" -> ToolExecution(knowledgeTools.deleteMemory(args))
                 "search_files" -> {
                     val query = args.getString("query")
                     val path = args.optString("path")
                     nativeFileManager.searchFiles(query, path)
-                        .fold({ ToolExecution(it.toSearchAgentText(query, path)) }, { throw it })
+                        .fold({ ToolExecution(it.toSearchAgentText(query, path, workspaceManager.displayName())) }, { throw it })
                 }
-                "global_search_files" -> globalSearchFiles(args.getString("query"))
+                "global_search_files" -> fileTools.globalSearchFiles(args.getString("query"))
                 "get_file_info" -> ToolExecution(nativeFileManager.fileInfo(args.getString("path")).getOrThrow())
                 "list_skill_files" -> ToolExecution(settings.listSkillFiles(args.getString("skill_id")).getOrThrow())
                 "read_skill_file" -> ToolExecution(settings.readSkillFile(args.getString("skill_id"), args.getString("path")).getOrThrow())
@@ -2296,174 +2086,13 @@ class OpenAiAgent(
                         allowShellFallback = true,
                     ).toJson(),
                 )
-                "list_email_accounts" -> ToolExecution(emailClient.accountsJson(settings.emailServers()))
-                "list_email_folders" -> ToolExecution(emailClient.listFolders(resolveEmailAccount(args)))
-                "list_emails" -> ToolExecution(
-                    emailClient.listMessages(
-                        account = resolveEmailAccount(args),
-                        folderName = args.optString("folder").ifBlank { "INBOX" },
-                        unreadOnly = args.optBoolean("unread_only", false),
-                        limit = args.optInt("limit", 30),
-                    ),
-                )
-                "read_email" -> ToolExecution(
-                    emailClient.readMessage(
-                        resolveEmailAccount(args),
-                        args.optString("folder").ifBlank { "INBOX" },
-                        args.getLong("uid"),
-                    ),
-                )
-                "set_email_flags" -> ToolExecution(
-                    emailClient.setFlags(
-                        resolveEmailAccount(args),
-                        args.optString("folder").ifBlank { "INBOX" },
-                        args.getLong("uid"),
-                        args.booleanOrNull("seen"),
-                        args.booleanOrNull("flagged"),
-                    ),
-                )
-                "download_email_attachment" -> ToolExecution(
-                    emailClient.downloadAttachment(
-                        resolveEmailAccount(args),
-                        args.optString("folder").ifBlank { "INBOX" },
-                        args.getLong("uid"),
-                        args.getInt("attachment_id"),
-                    ),
-                )
-                "record_email_attachment_scan" -> ToolExecution(
-                    emailClient.recordAttachmentScan(args.getString("attachment_token"), args.getBoolean("safe")),
-                )
-                "save_email_draft" -> ToolExecution(emailClient.saveDraft(resolveEmailAccount(args), emailComposeRequest(args)))
-                "send_email" -> ToolExecution(emailClient.send(resolveEmailAccount(args), emailComposeRequest(args)))
-                "list_ssh_servers" -> ToolExecution(sshExecutor.availableServers())
-                "ssh_exec" -> {
-                    val server = settings.resolveSshServer(args.getString("server_id"))
-                        ?: error("SSH server is missing or disabled: ${args.optString("server_id")}. Call list_ssh_servers and use a returned id.")
-                    val timeoutSeconds = args.optInt("timeout_seconds", server.timeoutSeconds).coerceIn(5, 600)
-                    val result = sshExecutor.execute(
-                        server = server,
-                        command = args.toolCommandArgument(),
-                        cwd = args.optString("cwd"),
-                        inputLines = args.optJSONArray("input_lines")?.let { array ->
-                            buildList {
-                                for (index in 0 until array.length()) add(array.optString(index))
-                            }
-                        }.orEmpty(),
-                        timeoutSeconds = timeoutSeconds,
-                    )
-                    if (result.ok) ToolExecution(result.message) else error(result.message)
-                }
-                "list_webdav_servers" -> ToolExecution(webDavClient.serversJson(settings.webDavServers().filter { it.enabled }))
-                "webdav_list" -> {
-                    val server = settings.resolveWebDavServer(args.getString("server_id"))
-                        ?: error("WebDAV server is missing or disabled: ${args.optString("server_id")}. Call list_webdav_servers and use a returned id.")
-                    val files = webDavClient.list(
-                        server = server,
-                        path = args.optString("path").ifBlank { server.initialPath.ifBlank { "/" } },
-                        depth = args.optInt("depth", 1).coerceIn(0, 2),
-                    )
-                    ToolExecution(webDavFilesJson(server, files).put("path", args.optString("path").ifBlank { server.initialPath.ifBlank { "/" } }).toString())
-                }
-                "webdav_search" -> {
-                    val server = settings.resolveWebDavServer(args.getString("server_id"))
-                        ?: error("WebDAV server is missing or disabled: ${args.optString("server_id")}. Call list_webdav_servers and use a returned id.")
-                    val files = webDavClient.search(
-                        server = server,
-                        query = args.getString("query"),
-                        basePath = args.optString("path").ifBlank { server.initialPath },
-                        limit = args.optInt("limit", 80).coerceIn(1, 200),
-                    )
-                    ToolExecution(webDavFilesJson(server, files).toString())
-                }
-                "webdav_download_to_workspace" -> {
-                    val server = settings.resolveWebDavServer(args.getString("server_id"))
-                        ?: error("WebDAV server is missing or disabled: ${args.optString("server_id")}. Call list_webdav_servers and use a returned id.")
-                    val bytes = webDavClient.download(server, args.getString("remote_path"))
-                    val message = nativeFileManager.writeBytes(args.getString("local_path"), bytes).getOrThrow()
-                    ToolExecution("$message\nDownloaded ${bytes.size} bytes from WebDAV.")
-                }
-                "webdav_upload_from_workspace" -> {
-                    val server = settings.resolveWebDavServer(args.getString("server_id"))
-                        ?: error("WebDAV server is missing or disabled: ${args.optString("server_id")}. Call list_webdav_servers and use a returned id.")
-                    val bytes = nativeFileManager.readBytes(args.getString("local_path")).getOrThrow()
-                    webDavClient.upload(server, args.getString("remote_path"), bytes)
-                    ToolExecution("Uploaded to WebDAV: ${server.name}:${args.getString("remote_path")}; ${bytes.size} bytes.")
-                }
-                "list_file_transfer_servers" -> ToolExecution(fileTransferClient.serversJson(settings.fileTransferServers().filter { it.enabled }))
-                "file_transfer_list" -> {
-                    val server = settings.resolveFileTransferServer(args.getString("server_id"))
-                        ?: error("File-transfer server is missing or disabled: ${args.optString("server_id")}. Call list_file_transfer_servers and use a returned id.")
-                    val path = args.optString("path").ifBlank { server.initialPath.ifBlank { "/" } }
-                    val files = fileTransferClient.list(server, path)
-                    ToolExecution(fileTransferFilesJson(server, files).put("path", path).toString())
-                }
-                "file_transfer_search" -> {
-                    val server = settings.resolveFileTransferServer(args.getString("server_id"))
-                        ?: error("File-transfer server is missing or disabled: ${args.optString("server_id")}. Call list_file_transfer_servers and use a returned id.")
-                    val files = fileTransferClient.search(
-                        server = server,
-                        query = args.getString("query"),
-                        basePath = args.optString("path").ifBlank { server.initialPath.ifBlank { "/" } },
-                        limit = args.optInt("limit", 80).coerceIn(1, 200),
-                    )
-                    ToolExecution(fileTransferFilesJson(server, files).toString())
-                }
-                "file_transfer_download_to_workspace" -> {
-                    val server = settings.resolveFileTransferServer(args.getString("server_id"))
-                        ?: error("File-transfer server is missing or disabled: ${args.optString("server_id")}. Call list_file_transfer_servers and use a returned id.")
-                    val bytes = fileTransferClient.download(server, args.getString("remote_path"))
-                    val message = nativeFileManager.writeBytes(args.getString("local_path"), bytes).getOrThrow()
-                    ToolExecution("$message\nDownloaded ${bytes.size} bytes from ${server.protocol.uppercase(Locale.US)}.")
-                }
-                "file_transfer_upload_from_workspace" -> {
-                    val server = settings.resolveFileTransferServer(args.getString("server_id"))
-                        ?: error("File-transfer server is missing or disabled: ${args.optString("server_id")}. Call list_file_transfer_servers and use a returned id.")
-                    val bytes = nativeFileManager.readBytes(args.getString("local_path")).getOrThrow()
-                    fileTransferClient.upload(server, args.getString("remote_path"), bytes)
-                    ToolExecution("Uploaded to ${server.protocol.uppercase(Locale.US)}: ${server.name}:${args.getString("remote_path")}; ${bytes.size} bytes.")
-                }
-                "export_backup" -> {
-                    val options = parseBackupOptions(args)
-                    val destination = args.optString("destination", "local").lowercase(Locale.US)
-                    if (destination == "webdav") {
-                        val server = settings.resolveWebDavServer(args.getString("server_id"))
-                            ?: error("WebDAV server is missing or disabled: ${args.optString("server_id")}. Call list_webdav_servers and use a returned id.")
-                        val remotePath = args.optString("remote_path").ifBlank { DEFAULT_WEBDAV_BACKUP_PATH }
-                        val bytes = backupManager.exportZip(options)
-                        webDavClient.upload(server, remotePath, bytes)
-                        ToolExecution(
-                            "Exported and uploaded the backup to WebDAV: ${server.name}:$remotePath; ${bytes.size} bytes.\n" +
-                                "When remote_path is omitted, the stable latest-backup path is overwritten so a later import does not need a timestamped name.",
-                        )
-                    } else {
-                        ToolExecution(backupManager.exportToDownloads(options))
-                    }
-                }
-                "import_backup" -> {
-                    val source = args.optString("source", "local").lowercase(Locale.US)
-                    val result = if (source == "webdav") {
-                        val server = settings.resolveWebDavServer(args.getString("server_id"))
-                            ?: error("WebDAV server is missing or disabled: ${args.optString("server_id")}. Call list_webdav_servers and use a returned id.")
-                        val remotePath = resolveWebDavBackupPath(server, args.optString("remote_path"))
-                        val bytes = webDavClient.download(server, remotePath)
-                        backupManager.importZip(bytes, "supplement")
-                    } else if (source == "download" || source == "global") {
-                        val path = args.optString("global_path").ifBlank { args.optString("local_path") }
-                        val bytes = globalFileManager.readBytes(path).getOrThrow()
-                        backupManager.importZip(bytes, "supplement")
-                    } else {
-                        val bytes = nativeFileManager.readBytes(args.getString("local_path")).getOrThrow()
-                        backupManager.importZip(bytes, "supplement")
-                    }
-                    configChangedHandler()
-                    ToolExecution("Imported the backup in supplement mode: $result")
-                }
+                in remoteTools.toolNames -> remoteTools.execute(call.name, args)
                 "get_mini_server_status" -> ToolExecution(miniServerManager.statusJson().toString())
-                "read_mini_server_logs" -> ToolExecution(readMiniServerLogs(args))
-                "manage_mini_server" -> ToolExecution(manageMiniServer(args))
+                "read_mini_server_logs" -> ToolExecution(automationTools.readMiniServerLogs(args))
+                "manage_mini_server" -> ToolExecution(automationTools.manageMiniServer(args))
                 "run_command" -> {
                     val command = args.toolCommandArgument()
-                    if (isFileSearchCommand(command)) {
+                    if (fileTools.isFileSearchCommand(command)) {
                         ToolExecution(
                             "ERROR: FILE_SEARCH_COMMAND_BLOCKED\n" +
                                 "Use search_files for file-name/path discovery instead of find, fd, or locate through run_command.\n" +
@@ -2481,7 +2110,7 @@ class OpenAiAgent(
                 }
                 "proot_command" -> {
                     val command = args.toolCommandArgument()
-                    if (isFileSearchCommand(command)) {
+                    if (fileTools.isFileSearchCommand(command)) {
                         ToolExecution(
                             "ERROR: FILE_SEARCH_COMMAND_BLOCKED\n" +
                                 "Use search_files for file-name/path discovery instead of find, fd, or locate through proot_command.",
@@ -2503,7 +2132,7 @@ class OpenAiAgent(
                 "web_search" -> ToolExecution(webAgent.search(args.getString("query"), args.optInt("limit", 6)))
                 "read_web_page" -> ToolExecution(webAgent.readPage(args.getString("url")))
                 "mark_web_sources" -> ToolExecution(webSourceMarkResult(args))
-                "manage_app_config" -> ToolExecution(manageAppConfig(args))
+                "manage_app_config" -> ToolExecution(configTools.manageAppConfig(args))
                 "run_sub_agents" -> ToolExecution(runSubAgents(conversationId, args, onStatus))
                 "ask_user" -> ToolExecution(
                     userQuestionHandler(parseUserQuestionRequest(conversationId, args)).toAgentJson(),
@@ -2930,30 +2559,6 @@ class OpenAiAgent(
         }
     }
 
-    private fun resolveWebDavBackupPath(server: WebDavServerConfig, requestedPath: String): String {
-        val explicit = requestedPath.trim()
-        if (explicit.isNotBlank()) return explicit
-        val files = runCatching { webDavClient.list(server, "/LyraCode", depth = 1) }.getOrDefault(emptyList())
-        val latest = files
-            .filter { !it.directory && it.path.endsWith(".zip", ignoreCase = true) }
-            .filter {
-                val name = it.path.substringAfterLast('/').lowercase(Locale.US)
-                "backup" in name || "lyra" in name
-            }
-        latest.firstOrNull { it.path.equals(DEFAULT_WEBDAV_BACKUP_PATH, ignoreCase = true) }?.let { return it.path }
-        return latest.maxWithOrNull(
-            compareBy<com.yukisoffd.lyracode.webdav.WebDavFile> { parseWebDavModifiedMillis(it.modified) }
-                .thenBy { it.path },
-        )?.path ?: DEFAULT_WEBDAV_BACKUP_PATH
-    }
-
-    private fun parseWebDavModifiedMillis(value: String): Long {
-        if (value.isBlank()) return 0L
-        return runCatching {
-            SimpleDateFormat("EEE, dd MMM yyyy HH:mm:ss z", Locale.US).parse(value)?.time ?: 0L
-        }.getOrDefault(0L)
-    }
-
     private suspend fun runSubAgents(parentConversationId: Long, args: JSONObject, onStatus: suspend (String) -> Unit = {}): String {
         require(!isSubAgentConversation(parentConversationId)) { "SUB_AGENT_RECURSION_BLOCKED: A sub-agent cannot delegate more sub-agents." }
         if (!settings.subAgentOrchestrationEnabled) return "ERROR: SUB_AGENT_DISABLED\nSub-agent orchestration is disabled by the user."
@@ -3300,539 +2905,6 @@ class OpenAiAgent(
             .toString()
     }
 
-    private suspend fun manageAppConfig(args: JSONObject): String {
-        val target = args.optString("target").trim().lowercase(Locale.US).replace("-", "_")
-        val action = args.optString("action").trim().lowercase(Locale.US).replace("-", "_")
-        require(target.isNotBlank()) { "target is required: all, mcp_server, ssh_server, email_server, webdav_server, file_transfer_server, skill, or agent_tool." }
-        require(action.isNotBlank()) { "action is required: list, add, update, enable, disable, or delete." }
-        val result = when (target) {
-            "all", "config", "configs", "inventory" -> {
-                require(action == "list") { "target=$target supports only action=list." }
-                configInventoryJson().toString()
-            }
-            "mcp", "mcp_server", "mcp_servers" -> manageMcpConfig(action, args)
-            "ssh", "ssh_server", "ssh_servers" -> manageSshConfig(action, args)
-            "email", "mail", "email_server", "email_servers", "imap", "smtp" -> manageEmailConfig(action, args)
-            "webdav", "webdav_server", "webdav_servers" -> manageWebDavConfig(action, args)
-            "file_transfer", "file_transfer_server", "file_transfer_servers", "ftp", "ftps", "sftp" -> manageFileTransferConfig(target, action, args)
-            "skill", "skills" -> manageSkillConfig(action, args)
-            "agent", "agent_tool", "tool", "tools" -> manageAgentToolConfig(action, args)
-            else -> error("Unknown configuration target: $target. Use target=all action=list to inspect supported targets.")
-        }
-        if (action != "list") {
-            configChangedHandler()
-        }
-        return result
-    }
-
-    private suspend fun manageMcpConfig(action: String, args: JSONObject): String {
-        if (action == "list") return configResult("mcp_servers", mcpServersJson()).toString()
-        val existing = resolveMcpServerForConfig(args.optString("id").ifBlank { args.optString("name") }.ifBlank { args.optString("url") })
-        when (action) {
-            "delete", "remove" -> {
-                val target = existing ?: error("MCP server to delete was not found. List configured servers and use an exact id or name.")
-                settings.deleteMcpServer(target.id)
-                return configResult("mcp_server_deleted", JSONObject().put("id", target.id).put("name", target.name)).toString()
-            }
-            "enable", "disable" -> {
-                val target = existing ?: error("MCP server to $action was not found. List configured servers and use an exact id or name.")
-                settings.setMcpServerEnabled(target.id, action == "enable")
-                return configResult("mcp_server_${action}d", mcpServerJson(target.copy(enabled = action == "enable"))).toString()
-            }
-        }
-
-                require(action in setOf("add", "create", "update", "modify", "upsert")) { "MCP does not support action=$action." }
-        val rawJson = args.optString("raw_json").ifBlank { existing?.rawJson.orEmpty() }
-        val parsed = parseMcpRawJson(rawJson)
-        val url = args.optString("url")
-            .ifBlank { args.optString("base_url") }
-            .ifBlank { parsed?.url.orEmpty() }
-            .ifBlank { existing?.url.orEmpty() }
-            .trim()
-                require(url.isNotBlank()) { "MCP url is required. If authentication data is missing, ask the user for the key or complete raw_json." }
-        val name = args.optString("name")
-            .ifBlank { parsed?.name.orEmpty() }
-            .ifBlank { existing?.name.orEmpty() }
-            .ifBlank { "MCP Server" }
-        val authKey = args.optString("auth_key")
-            .ifBlank { args.optString("api_key") }
-            .ifBlank { args.optString("key") }
-            .ifBlank { parsed?.authKey.orEmpty() }
-            .ifBlank { existing?.authKey.orEmpty() }
-        val transport = normalizeMcpTransport(
-            args.optString("transport")
-                .ifBlank { parsed?.transport.orEmpty() }
-                .ifBlank { existing?.transport.orEmpty() },
-        )
-        val timeout = args.optInt("timeout_seconds", existing?.timeoutSeconds ?: 30).coerceIn(5, 300)
-        val enabled = if (args.has("enabled")) args.optBoolean("enabled") else existing?.enabled ?: true
-        val server = McpServerConfig(
-            id = existing?.id ?: args.optString("id").ifBlank { AppSettings.newId() },
-            name = name,
-            url = url,
-            authKey = authKey,
-            transport = transport,
-            timeoutSeconds = timeout,
-            enabled = enabled,
-            rawJson = buildMcpRawJson(rawJson, name, url, authKey, transport),
-            tools = existing?.tools.orEmpty(),
-        )
-        settings.upsertMcpServer(server)
-        val refresh = if (enabled) {
-            runCatching { mcpClientManager.testAndRefreshTools(server).getOrThrow() }
-        } else {
-            Result.success(server.tools)
-        }
-        val saved = settings.mcpServers().firstOrNull { it.id == server.id } ?: server
-        return configResult(
-            "mcp_server_saved",
-            JSONObject()
-                .put("server", mcpServerJson(saved))
-                .put("tools_count", saved.tools.size)
-                .put("refresh_ok", refresh.isSuccess)
-                    .put("message", refresh.exceptionOrNull()?.message.orEmpty().ifBlank { "MCP server saved and tools refreshed." }),
-        ).toString()
-    }
-
-    private fun manageSshConfig(action: String, args: JSONObject): String {
-        if (action == "list") return configResult("ssh_servers", sshServersJson()).toString()
-        val existing = resolveSshServerForConfig(args.optString("id").ifBlank { args.optString("host") }.ifBlank { args.optString("name") })
-        when (action) {
-            "delete", "remove" -> {
-                val target = existing ?: error("SSH server to delete was not found. List configured servers and use an exact id or name.")
-                settings.deleteSshServer(target.id)
-                return configResult("ssh_server_deleted", JSONObject().put("id", target.id).put("host", target.host)).toString()
-            }
-            "enable", "disable" -> {
-                val target = existing ?: error("SSH server to $action was not found. List configured servers and use an exact id or name.")
-                settings.setSshServerEnabled(target.id, action == "enable")
-                return configResult("ssh_server_${action}d", sshServerJson(target.copy(enabled = action == "enable"))).toString()
-            }
-        }
-                require(action in setOf("add", "create", "update", "modify", "upsert")) { "SSH does not support action=$action." }
-        val host = args.optString("host").ifBlank { existing?.host.orEmpty() }.trim()
-        val username = args.optString("username").ifBlank { args.optString("user") }.ifBlank { existing?.username.orEmpty() }.trim()
-                require(host.isNotBlank()) { "SSH host is required." }
-                require(username.isNotBlank()) { "SSH username is required." }
-        val authType = when (args.optString("auth_type").ifBlank { existing?.authType.orEmpty() }.lowercase(Locale.US)) {
-            "key", "private_key", "ssh_key" -> AppSettings.SSH_AUTH_KEY
-            else -> AppSettings.SSH_AUTH_PASSWORD
-        }
-        val server = SshServerConfig(
-            id = existing?.id ?: args.optString("id").ifBlank { AppSettings.newId() },
-            name = args.optString("name").ifBlank { existing?.name.orEmpty() }.ifBlank { host },
-            host = host,
-            port = args.optInt("port", existing?.port ?: 22).coerceIn(1, 65535),
-            username = username,
-            authType = authType,
-            password = args.optString("password").ifBlank { existing?.password.orEmpty() },
-            privateKey = args.optString("private_key").ifBlank { existing?.privateKey.orEmpty() },
-            passphrase = args.optString("passphrase").ifBlank { existing?.passphrase.orEmpty() },
-            timeoutSeconds = args.optInt("timeout_seconds", existing?.timeoutSeconds ?: 60).coerceIn(5, 600),
-            enabled = if (args.has("enabled")) args.optBoolean("enabled") else existing?.enabled ?: true,
-        )
-                require(server.authType != AppSettings.SSH_AUTH_PASSWORD || server.password.isNotBlank()) { "Password authentication requires password. Ask the user if it was not provided." }
-                require(server.authType != AppSettings.SSH_AUTH_KEY || server.privateKey.isNotBlank()) { "Key authentication requires private_key. Ask the user if it was not provided." }
-        settings.upsertSshServer(server)
-        return configResult("ssh_server_saved", sshServerJson(server)).toString()
-    }
-
-    private fun manageEmailConfig(action: String, args: JSONObject): String {
-        if (action == "list") return configResult("email_servers", emailServersJson()).toString()
-        val key = args.optString("id").ifBlank { args.optString("email_address") }.ifBlank { args.optString("name") }
-        val existing = settings.emailServers().firstOrNull {
-            it.id == key || it.name == key || it.stableId.equals(key, ignoreCase = true)
-        }
-        when (action) {
-            "delete", "remove" -> {
-                val target = existing ?: error("Email account to delete was not found. List configured accounts and use an exact id or address.")
-                settings.deleteEmailServer(target.id)
-                return configResult("email_server_deleted", JSONObject().put("id", target.id).put("email_address", target.emailAddress)).toString()
-            }
-            "enable", "disable" -> {
-                val target = existing ?: error("Email account to $action was not found. List configured accounts and use an exact id or address.")
-                settings.setEmailServerEnabled(target.id, action == "enable")
-                return configResult("email_server_${action}d", emailServerJson(target.copy(enabled = action == "enable"))).toString()
-            }
-        }
-        require(action in setOf("add", "create", "update", "modify", "upsert")) { "Email configuration does not support action=$action." }
-        val address = args.optString("email_address").ifBlank { existing?.emailAddress.orEmpty() }.trim()
-        val username = args.optString("username").ifBlank { existing?.username.orEmpty() }.ifBlank { address }.trim()
-        val password = args.optString("password").ifBlank { existing?.password.orEmpty() }
-        val imapHost = args.optString("imap_host").ifBlank { existing?.imapHost.orEmpty() }.trim()
-        val smtpHost = args.optString("smtp_host").ifBlank { existing?.smtpHost.orEmpty() }.trim()
-        require(address.contains('@') && address.substringAfter('@').contains('.')) { "A valid email_address is required." }
-        require(username.isNotBlank() && password.isNotBlank()) { "Email username and password/app password are required. Ask the user; never invent credentials." }
-        require(imapHost.isNotBlank() && smtpHost.isNotBlank()) { "Both imap_host and smtp_host are required." }
-        val server = EmailServerConfig(
-            id = existing?.id ?: args.optString("id").ifBlank { AppSettings.newId() },
-            name = args.optString("name").ifBlank { existing?.name.orEmpty() }.ifBlank { address },
-            emailAddress = address,
-            username = username,
-            password = password,
-            imapHost = imapHost,
-            imapPort = args.optInt("imap_port", existing?.imapPort ?: 993).coerceIn(1, 65535),
-            imapSecurity = AppSettings.normalizeEmailSecurity(args.optString("imap_security").ifBlank { existing?.imapSecurity.orEmpty() }),
-            smtpHost = smtpHost,
-            smtpPort = args.optInt("smtp_port", existing?.smtpPort ?: 465).coerceIn(1, 65535),
-            smtpSecurity = AppSettings.normalizeEmailSecurity(args.optString("smtp_security").ifBlank { existing?.smtpSecurity.orEmpty() }),
-            enabled = if (args.has("enabled")) args.optBoolean("enabled") else existing?.enabled ?: true,
-        )
-        settings.upsertEmailServer(server)
-        return configResult("email_server_saved", emailServerJson(server)).toString()
-    }
-
-    private fun manageWebDavConfig(action: String, args: JSONObject): String {
-        if (action == "list") return configResult("webdav_servers", webDavServersJson()).toString()
-        val existing = resolveWebDavServerForConfig(
-            args.optString("id")
-                .ifBlank { args.optString("url") }
-                .ifBlank { args.optString("name") },
-        )
-        when (action) {
-            "delete", "remove" -> {
-                val target = existing ?: error("WebDAV server to delete was not found. List configured servers and use an exact id or name.")
-                settings.deleteWebDavServer(target.id)
-                return configResult("webdav_server_deleted", JSONObject().put("id", target.id).put("name", target.name)).toString()
-            }
-            "enable", "disable" -> {
-                val target = existing ?: error("WebDAV server to $action was not found. List configured servers and use an exact id or name.")
-                settings.setWebDavServerEnabled(target.id, action == "enable")
-                return configResult("webdav_server_${action}d", webDavServerJson(target.copy(enabled = action == "enable"))).toString()
-            }
-        }
-                require(action in setOf("add", "create", "update", "modify", "upsert")) { "WebDAV does not support action=$action." }
-        val url = args.optString("url").ifBlank { args.optString("base_url") }.ifBlank { existing?.url.orEmpty() }.trim()
-                require(url.isNotBlank()) { "WebDAV url is required." }
-                require(url.startsWith("http://", true) || url.startsWith("https://", true)) { "WebDAV url must use http:// or https://." }
-        val server = WebDavServerConfig(
-            id = existing?.id ?: args.optString("id").ifBlank { AppSettings.newId() },
-            name = args.optString("name").ifBlank { existing?.name.orEmpty() }.ifBlank { runCatching { URI(url).host }.getOrNull().orEmpty().ifBlank { "WebDAV" } },
-            url = url,
-            username = args.optString("username").ifBlank { args.optString("user") }.ifBlank { existing?.username.orEmpty() },
-            password = args.optString("password").ifBlank { existing?.password.orEmpty() },
-            userAgent = args.optString("user_agent").ifBlank { existing?.userAgent.orEmpty() },
-            initialPath = args.optString("initial_path").ifBlank { args.optString("path") }.ifBlank { existing?.initialPath.orEmpty() }.ifBlank { "/" },
-            note = args.optString("note").ifBlank { existing?.note.orEmpty() },
-            trustAllCertificates = if (args.has("trust_all_certificates")) args.optBoolean("trust_all_certificates") else existing?.trustAllCertificates ?: false,
-            multiThread = if (args.has("multi_thread")) args.optBoolean("multi_thread") else existing?.multiThread ?: true,
-            hideAddressInDrawer = if (args.has("hide_address")) args.optBoolean("hide_address") else existing?.hideAddressInDrawer ?: false,
-            enabled = if (args.has("enabled")) args.optBoolean("enabled") else existing?.enabled ?: true,
-        )
-        settings.upsertWebDavServer(server)
-        val test = if (server.enabled) webDavClient.test(server) else Result.success(emptyList())
-        return configResult(
-            "webdav_server_saved",
-            JSONObject()
-                .put("server", webDavServerJson(server))
-                .put("test_ok", test.isSuccess)
-                    .put("message", test.exceptionOrNull()?.message.orEmpty().ifBlank { if (server.url.startsWith("http://", true)) "Saved. Warning: plain HTTP is insecure." else "WebDAV saved and connection test passed." }),
-        ).toString()
-    }
-
-    private fun manageFileTransferConfig(target: String, action: String, args: JSONObject): String {
-        if (action == "list") return configResult("file_transfer_servers", fileTransferServersJson()).toString()
-        val protocolHint = when (target) {
-            "ftp", "ftps", "sftp" -> target
-            else -> ""
-        }
-        val existing = resolveFileTransferServerForConfig(
-            args.optString("id")
-                .ifBlank { args.optString("host") }
-                .ifBlank { args.optString("name") },
-        )
-        when (action) {
-            "delete", "remove" -> {
-                val targetServer = existing ?: error("File-transfer server to delete was not found. List configured servers and use an exact id or name.")
-                settings.deleteFileTransferServer(targetServer.id)
-                return configResult("file_transfer_server_deleted", JSONObject().put("id", targetServer.id).put("name", targetServer.name)).toString()
-            }
-            "enable", "disable" -> {
-                val targetServer = existing ?: error("File-transfer server to $action was not found. List configured servers and use an exact id or name.")
-                settings.setFileTransferServerEnabled(targetServer.id, action == "enable")
-                return configResult("file_transfer_server_${action}d", fileTransferServerJson(targetServer.copy(enabled = action == "enable"))).toString()
-            }
-        }
-                require(action in setOf("add", "create", "update", "modify", "upsert")) { "File-transfer server does not support action=$action." }
-        val protocol = AppSettings.normalizeFileTransferProtocol(
-            args.optString("protocol")
-                .ifBlank { protocolHint }
-                .ifBlank { existing?.protocol.orEmpty() }
-                .ifBlank { AppSettings.FILE_TRANSFER_SFTP },
-        )
-        val host = args.optString("host").ifBlank { args.optString("url") }.ifBlank { existing?.host.orEmpty() }.trim()
-                require(host.isNotBlank()) { "File-transfer server host is required." }
-        val username = args.optString("username").ifBlank { args.optString("user") }.ifBlank { existing?.username.orEmpty() }.trim()
-                if (protocol == AppSettings.FILE_TRANSFER_SFTP) require(username.isNotBlank()) { "SFTP requires username. Ask the user if it was not provided." }
-        val usePrivateKey = if (args.has("use_private_key")) args.optBoolean("use_private_key") else existing?.usePrivateKey ?: false
-        val server = FileTransferServerConfig(
-            id = existing?.id ?: args.optString("id").ifBlank { AppSettings.newId() },
-            name = args.optString("name").ifBlank { existing?.name.orEmpty() }.ifBlank { "${protocol.uppercase(Locale.US)} $host" },
-            protocol = protocol,
-            host = host,
-            port = args.optInt("port", existing?.port ?: AppSettings.defaultFileTransferPort(protocol)).coerceIn(1, 65535),
-            username = username.ifBlank { if (protocol == AppSettings.FILE_TRANSFER_SFTP) "" else "anonymous" },
-            password = args.optString("password").ifBlank { existing?.password.orEmpty() },
-            usePrivateKey = usePrivateKey,
-            privateKey = args.optString("private_key").ifBlank { existing?.privateKey.orEmpty() },
-            passphrase = args.optString("passphrase").ifBlank { existing?.passphrase.orEmpty() },
-            initialPath = args.optString("initial_path").ifBlank { args.optString("path") }.ifBlank { existing?.initialPath.orEmpty() }.ifBlank { "/" },
-            note = args.optString("note").ifBlank { existing?.note.orEmpty() },
-            encoding = args.optString("encoding").ifBlank { existing?.encoding.orEmpty() }.ifBlank { "UTF-8" },
-            passiveMode = if (args.has("passive_mode")) args.optBoolean("passive_mode") else existing?.passiveMode ?: true,
-            explicitFtps = if (args.has("explicit_ftps")) args.optBoolean("explicit_ftps") else existing?.explicitFtps ?: true,
-            multiThread = if (args.has("multi_thread")) args.optBoolean("multi_thread") else existing?.multiThread ?: true,
-            syncPermissions = if (args.has("sync_permissions")) args.optBoolean("sync_permissions") else existing?.syncPermissions ?: false,
-            hideAddressInDrawer = if (args.has("hide_address")) args.optBoolean("hide_address") else existing?.hideAddressInDrawer ?: false,
-            enabled = if (args.has("enabled")) args.optBoolean("enabled") else existing?.enabled ?: true,
-        )
-                require(!server.usePrivateKey || server.privateKey.isNotBlank()) { "Key authentication requires private_key. Ask the user if it was not provided." }
-        settings.upsertFileTransferServer(server)
-        val test = if (server.enabled) fileTransferClient.test(server) else Result.success(emptyList())
-        return configResult(
-            "file_transfer_server_saved",
-            JSONObject()
-                .put("server", fileTransferServerJson(server))
-                .put("test_ok", test.isSuccess)
-                .put("message", test.exceptionOrNull()?.message.orEmpty().ifBlank {
-                        if (server.protocol == AppSettings.FILE_TRANSFER_FTP) "Saved. Warning: FTP is plaintext; prefer SFTP or FTPS." else "File-transfer server saved and connection test passed."
-                }),
-        ).toString()
-    }
-
-    private fun manageSkillConfig(action: String, args: JSONObject): String {
-        if (action == "list") return configResult("skills", skillsJson()).toString()
-        val existing = resolveSkillForConfig(args.optString("id").ifBlank { args.optString("name") })
-        when (action) {
-            "add", "create", "install", "import" -> {
-                val url = args.optString("zip_url").ifBlank { args.optString("url") }.trim()
-                require(url.isNotBlank()) { "Installing a Skill requires zip_url. If the user provided a web page, read it and locate the actual zip URL." }
-                val download = downloadBytes(url)
-                val skill = settings.importSkillZipBytes(args.optString("name").ifBlank { download.first }, download.second).getOrThrow()
-                args.optString("description").takeIf { it.isNotBlank() }?.let { settings.updateSkillMeta(skill.id, description = it) }
-                return configResult("skill_installed", skillJson(settings.installedSkills().firstOrNull { it.id == skill.id } ?: skill)).toString()
-            }
-            "delete", "remove", "uninstall" -> {
-                val target = existing ?: error("Skill to delete was not found. List configured Skills and use an exact id or name.")
-                settings.deleteSkill(target.id)
-                return configResult("skill_deleted", JSONObject().put("id", target.id).put("name", target.name)).toString()
-            }
-            "enable", "disable" -> {
-                val target = existing ?: error("Skill to $action was not found. List configured Skills and use an exact id or name.")
-                settings.setSkillEnabled(target.id, action == "enable")
-                return configResult("skill_${action}d", skillJson(target.copy(enabled = action == "enable"))).toString()
-            }
-            "update", "modify", "rename" -> {
-                val target = existing ?: error("Skill to update was not found. List configured Skills and use an exact id or name.")
-                settings.updateSkillMeta(target.id, args.optString("name").ifBlank { null }, args.optString("description").ifBlank { null })
-                if (args.has("enabled")) settings.setSkillEnabled(target.id, args.optBoolean("enabled"))
-                val updated = settings.installedSkills().firstOrNull { it.id == target.id } ?: target
-                return configResult("skill_updated", skillJson(updated)).toString()
-            }
-            else -> error("Skill does not support action=$action.")
-        }
-    }
-
-    private fun manageAgentToolConfig(action: String, args: JSONObject): String {
-        if (action == "list") return configResult("agent_tools", agentToolsJson()).toString()
-        val toolName = args.optString("tool_name").ifBlank { args.optString("name") }.ifBlank { args.optString("id") }.trim()
-        require(toolName.isNotBlank()) { "Managing an Agent tool requires tool_name." }
-        require(toolName != "manage_app_config") { "manage_app_config is protected and cannot be disabled or deleted." }
-        return when (action) {
-            "enable" -> {
-                settings.setToolEnabled(toolName, true)
-                configResult("agent_tool_enabled", JSONObject().put("tool_name", toolName)).toString()
-            }
-            "disable" -> {
-                settings.setToolEnabled(toolName, false)
-                configResult("agent_tool_disabled", JSONObject().put("tool_name", toolName)).toString()
-            }
-            "update", "modify" -> {
-                require(args.has("enabled")) { "Agent tools can only be updated with enabled=true or enabled=false." }
-                settings.setToolEnabled(toolName, args.optBoolean("enabled"))
-                configResult("agent_tool_updated", JSONObject().put("tool_name", toolName).put("enabled", args.optBoolean("enabled"))).toString()
-            }
-            "delete", "remove" -> error("Built-in Agent tools cannot be deleted; enable or disable them instead.")
-            else -> error("Agent tools do not support action=$action.")
-        }
-    }
-
-    private fun configResult(type: String, payload: Any): JSONObject {
-        return JSONObject()
-            .put("schema", "lyra_config_management_result_v1")
-            .put("type", type)
-            .put("payload", payload)
-    }
-
-    private fun configInventoryJson(): JSONObject {
-        return configResult(
-            "config_inventory",
-            JSONObject()
-                .put("agent_tools", agentToolsJson())
-                .put("mcp_servers", mcpServersJson())
-                .put("ssh_servers", sshServersJson())
-                .put("email_servers", emailServersJson())
-                .put("webdav_servers", webDavServersJson())
-                .put("file_transfer_servers", fileTransferServersJson())
-                .put("skills", skillsJson())
-                .put("disabled_summary", disabledConfigSummaryJson())
-                .put("instruction", "Before enabling an item, confirm its id, name, or tool_name from disabled_summary or the matching list. Use the corresponding target for MCP, SSH, email, WebDAV, file-transfer, Skill, or Agent-tool configuration."),
-        )
-    }
-
-    private fun disabledConfigSummaryJson(): JSONObject {
-        val disabledTools = settings.disabledTools()
-        val mcpServers = settings.mcpServers()
-        return JSONObject()
-            .put("agent_tools", JSONArray().also { array ->
-                agentToolNamesForConfig().filter { it != "manage_app_config" && it in disabledTools }.sorted().forEach { array.put(it) }
-            })
-            .put("mcp_servers", JSONArray().also { array ->
-                mcpServers.filterNot { it.enabled }.forEach { array.put(JSONObject().put("id", it.id).put("name", it.name).put("url", it.url)) }
-            })
-            .put("mcp_tools_unavailable", JSONArray().also { array ->
-                mcpServers.forEach { server ->
-                    server.tools.forEach { tool ->
-                        val functionName = settings.mcpToolFunctionName(server, tool)
-                        if (!server.enabled || functionName in disabledTools) {
-                            array.put(
-                                JSONObject()
-                                    .put("tool_name", functionName)
-                                    .put("server_id", server.id)
-                                    .put("server_name", server.name)
-                                    .put("mcp_tool", tool.name)
-                                    .put("server_enabled", server.enabled)
-                                    .put("tool_enabled", functionName !in disabledTools),
-                            )
-                        }
-                    }
-                }
-            })
-            .put("ssh_servers", JSONArray().also { array ->
-                settings.sshServers().filterNot { it.enabled }.forEach { array.put(JSONObject().put("id", it.id).put("name", it.name).put("host", it.host)) }
-            })
-            .put("email_servers", JSONArray().also { array ->
-                settings.emailServers().filterNot { it.enabled }.forEach {
-                    array.put(JSONObject().put("id", it.id).put("name", it.name).put("email_address", it.emailAddress))
-                }
-            })
-            .put("webdav_servers", JSONArray().also { array ->
-                settings.webDavServers().filterNot { it.enabled }.forEach { array.put(JSONObject().put("id", it.id).put("name", it.name).put("url", it.url)) }
-            })
-            .put("file_transfer_servers", JSONArray().also { array ->
-                settings.fileTransferServers().filterNot { it.enabled }.forEach {
-                    array.put(JSONObject().put("id", it.id).put("name", it.name).put("protocol", it.protocol).put("host", it.host))
-                }
-            })
-            .put("skills", JSONArray().also { array ->
-                settings.installedSkills().filterNot { it.enabled }.forEach { array.put(JSONObject().put("id", it.id).put("name", it.name).put("description", it.description)) }
-            })
-    }
-
-    private data class ParsedMcpRawConfig(
-        val name: String,
-        val url: String,
-        val authKey: String,
-        val transport: String,
-        val serverKey: String,
-    )
-
-    private fun parseMcpRawJson(rawJson: String): ParsedMcpRawConfig? = runCatching {
-        if (rawJson.isBlank()) return@runCatching null
-        val root = JSONObject(rawJson)
-        val servers = root.optJSONObject("mcpServers")
-        val serverKey = servers?.keys()?.asSequence()?.firstOrNull().orEmpty()
-        val node = if (serverKey.isNotBlank()) servers?.optJSONObject(serverKey) else root
-        node ?: return@runCatching null
-        val headers = node.optJSONObject("headers") ?: root.optJSONObject("headers")
-        val auth = headers?.optString("Authorization").orEmpty().removePrefix("Bearer ").trim()
-        val rawType = node.optString("type").ifBlank { node.optString("transport") }
-        ParsedMcpRawConfig(
-            name = node.optString("name").ifBlank { serverKey.ifBlank { root.optString("name") } },
-            url = node.optString("baseUrl").ifBlank { node.optString("url").ifBlank { root.optString("baseUrl").ifBlank { root.optString("url") } } },
-            authKey = auth,
-            transport = normalizeMcpTransport(rawType),
-            serverKey = serverKey.ifBlank { node.optString("id").ifBlank { "mcp_server" } },
-        )
-    }.getOrNull()
-
-    private fun buildMcpRawJson(rawJson: String, name: String, url: String, authKey: String, transport: String): String {
-        val parsed = parseMcpRawJson(rawJson)
-        val serverKey = parsed?.serverKey?.takeIf { it.isNotBlank() } ?: configKeyPart(name).ifBlank { "mcp_server" }
-        val root = runCatching { JSONObject(rawJson.ifBlank { "{}" }) }.getOrDefault(JSONObject())
-        val servers = root.optJSONObject("mcpServers") ?: JSONObject()
-        val node = servers.optJSONObject(serverKey) ?: JSONObject()
-        node.put("type", if (transport == AppSettings.MCP_TRANSPORT_SSE) "sse" else "streamableHttp")
-        node.put("name", name)
-        node.put("baseUrl", url)
-        val headers = node.optJSONObject("headers") ?: JSONObject()
-        if (authKey.isNotBlank()) {
-            headers.put("Authorization", if (authKey.startsWith("Bearer ", ignoreCase = true)) authKey else "Bearer $authKey")
-        }
-        node.put("headers", headers)
-        servers.put(serverKey, node)
-        root.put("mcpServers", servers)
-        if (!root.has("protocolVersion")) root.put("protocolVersion", "2025-06-18")
-        return root.toString()
-    }
-
-    private fun normalizeMcpTransport(raw: String): String {
-        return when (raw.trim().lowercase(Locale.US)) {
-            "sse" -> AppSettings.MCP_TRANSPORT_SSE
-            else -> AppSettings.MCP_TRANSPORT_STREAMABLE_HTTP
-        }
-    }
-
-    private fun configKeyPart(value: String): String {
-        return value.lowercase(Locale.US)
-            .replace(Regex("[^a-z0-9_]+"), "_")
-            .trim('_')
-    }
-
-    private fun resolveMcpServerForConfig(identifier: String): McpServerConfig? {
-        val clean = identifier.trim()
-        if (clean.isBlank()) return null
-        return settings.mcpServers().firstOrNull { it.id == clean || it.name == clean || it.url == clean }
-    }
-
-    private fun resolveSshServerForConfig(identifier: String): SshServerConfig? {
-        val clean = identifier.trim()
-        if (clean.isBlank()) return null
-        return settings.sshServers().firstOrNull { it.id == clean || it.stableId == clean || it.host == clean || it.name == clean }
-    }
-
-    private fun resolveWebDavServerForConfig(identifier: String): WebDavServerConfig? {
-        val clean = identifier.trim().trimEnd('/')
-        if (clean.isBlank()) return null
-        return settings.webDavServers().firstOrNull {
-            it.id == clean || it.name == clean || it.stableId == clean || it.url.trimEnd('/') == clean
-        }
-    }
-
-    private fun resolveSkillForConfig(identifier: String): SkillPack? {
-        val clean = identifier.trim()
-        if (clean.isBlank()) return null
-        return settings.installedSkills().firstOrNull { it.id == clean || it.name == clean }
-    }
-
-    private fun downloadBytes(url: String): Pair<String, ByteArray> {
-        require(url.startsWith("http://", true) || url.startsWith("https://", true)) { "Download URL must use http:// or https://." }
-        val request = Request.Builder().url(url).get().build()
-        client.newCall(request).execute().use { response ->
-            val body = response.body ?: error("Download response has no body.")
-            if (!response.isSuccessful) error("Download failed with HTTP ${response.code}: ${body.string().take(500)}")
-            val bytes = body.bytes()
-            require(bytes.isNotEmpty()) { "Downloaded file is empty." }
-            require(bytes.size <= 16 * 1024 * 1024) { "Downloaded file exceeds the 16 MB Skill-import limit." }
-            val fileName = response.header("Content-Disposition")
-                ?.substringAfter("filename=", "")
-                ?.trim('"', '\'')
-                ?.takeIf { it.isNotBlank() }
-                ?: runCatching { URI(url).path.substringAfterLast('/') }.getOrNull().orEmpty().ifBlank { "Skill.zip" }
-            return fileName to bytes
-        }
-    }
-
     private suspend fun downloadFile(args: JSONObject): ToolExecution {
         val url = args.getString("url").trim()
         require(url.startsWith("http://", true) || url.startsWith("https://", true)) {
@@ -3884,631 +2956,6 @@ class OpenAiAgent(
                 .put("sha256", result.sha256)
                 .toString(),
         )
-    }
-
-    private fun mcpServersJson(): JSONArray = JSONArray().also { array ->
-        settings.mcpServers().forEach { array.put(mcpServerJson(it)) }
-    }
-
-    private fun mcpServerJson(server: McpServerConfig): JSONObject = JSONObject()
-        .put("id", server.id)
-        .put("name", server.name)
-        .put("url", server.url)
-        .put("transport", server.transport)
-        .put("timeout_seconds", server.timeoutSeconds)
-        .put("enabled", server.enabled)
-        .put("tools", JSONArray().also { tools ->
-            val disabled = settings.disabledTools()
-            server.tools.forEach { tool ->
-                val functionName = settings.mcpToolFunctionName(server, tool)
-                tools.put(
-                    JSONObject()
-                        .put("name", tool.name)
-                        .put("function_name", functionName)
-                        .put("description", tool.description)
-                        .put("enabled", server.enabled && functionName !in disabled)
-                        .put("server_enabled", server.enabled)
-                        .put("tool_enabled", functionName !in disabled),
-                )
-            }
-        })
-
-    private fun sshServersJson(): JSONArray = JSONArray().also { array ->
-        settings.sshServers().forEach { array.put(sshServerJson(it)) }
-    }
-
-    private fun sshServerJson(server: SshServerConfig): JSONObject = JSONObject()
-        .put("id", server.id)
-        .put("stable_id", server.stableId)
-        .put("name", server.name)
-        .put("host", server.host)
-        .put("port", server.port)
-        .put("username", server.username)
-        .put("auth_type", server.authType)
-        .put("timeout_seconds", server.timeoutSeconds)
-        .put("enabled", server.enabled)
-        .put("has_password", server.password.isNotBlank())
-        .put("has_private_key", server.privateKey.isNotBlank())
-
-    private fun emailServersJson(): JSONArray = JSONArray().also { array ->
-        settings.emailServers().forEach { array.put(emailServerJson(it)) }
-    }
-
-    private fun emailServerJson(server: EmailServerConfig): JSONObject = JSONObject()
-        .put("id", server.id)
-        .put("stable_id", server.stableId)
-        .put("name", server.name)
-        .put("email_address", server.emailAddress)
-        .put("username", server.username)
-        .put("imap_host", server.imapHost)
-        .put("imap_port", server.imapPort)
-        .put("imap_security", server.imapSecurity)
-        .put("smtp_host", server.smtpHost)
-        .put("smtp_port", server.smtpPort)
-        .put("smtp_security", server.smtpSecurity)
-        .put("enabled", server.enabled)
-        .put("has_password", server.password.isNotBlank())
-
-    private fun webDavServersJson(): JSONArray = JSONArray().also { array ->
-        settings.webDavServers().forEach { array.put(webDavServerJson(it)) }
-    }
-
-    private fun webDavServerJson(server: WebDavServerConfig): JSONObject = JSONObject()
-        .put("id", server.id)
-        .put("stable_id", server.stableId)
-        .put("name", server.name)
-        .put("url", server.url)
-        .put("username", server.username)
-        .put("initial_path", server.initialPath)
-        .put("note", server.note)
-        .put("enabled", server.enabled)
-        .put("trust_all_certificates", server.trustAllCertificates)
-        .put("multi_thread", server.multiThread)
-        .put("hide_address", server.hideAddressInDrawer)
-        .put("has_password", server.password.isNotBlank())
-
-    private fun webDavFilesJson(server: WebDavServerConfig, files: List<com.yukisoffd.lyracode.webdav.WebDavFile>): JSONObject = JSONObject()
-        .put("schema", "lyra_webdav_files_v1")
-        .put("server_id", server.id)
-        .put("server_name", server.name)
-        .put("files", JSONArray().also { array ->
-            files.forEach { file ->
-                array.put(
-                    JSONObject()
-                        .put("path", file.path)
-                        .put("directory", file.directory)
-                        .put("size", file.size)
-                        .put("modified", file.modified),
-                )
-            }
-        })
-
-    private fun fileTransferServersJson(): JSONArray = JSONArray().also { array ->
-        settings.fileTransferServers().forEach { array.put(fileTransferServerJson(it)) }
-    }
-
-    private fun resolveFileTransferServerForConfig(key: String): FileTransferServerConfig? {
-        val clean = key.trim()
-        if (clean.isBlank()) return null
-        return settings.fileTransferServers().firstOrNull {
-            it.id == clean ||
-                it.name.equals(clean, ignoreCase = true) ||
-                it.host.equals(clean, ignoreCase = true) ||
-                it.stableId.equals(clean, ignoreCase = true)
-        }
-    }
-
-    private fun fileTransferServerJson(server: FileTransferServerConfig): JSONObject = JSONObject()
-        .put("id", server.id)
-        .put("stable_id", server.stableId)
-        .put("name", server.name)
-        .put("protocol", server.protocol)
-        .put("host", server.host)
-        .put("port", server.port)
-        .put("username", server.username)
-        .put("initial_path", server.initialPath)
-        .put("note", server.note)
-        .put("encoding", server.encoding)
-        .put("enabled", server.enabled)
-        .put("use_private_key", server.usePrivateKey)
-        .put("passive_mode", server.passiveMode)
-        .put("explicit_ftps", server.explicitFtps)
-        .put("multi_thread", server.multiThread)
-        .put("sync_permissions", server.syncPermissions)
-        .put("hide_address", server.hideAddressInDrawer)
-        .put("has_password", server.password.isNotBlank())
-        .put("has_private_key", server.privateKey.isNotBlank())
-
-    private fun fileTransferFilesJson(server: FileTransferServerConfig, files: List<com.yukisoffd.lyracode.filetransfer.FileTransferFile>): JSONObject = JSONObject()
-        .put("schema", "lyra_file_transfer_files_v1")
-        .put("server_id", server.id)
-        .put("server_name", server.name)
-        .put("protocol", server.protocol)
-        .put("files", JSONArray().also { array ->
-            files.forEach { file ->
-                array.put(
-                    JSONObject()
-                        .put("path", file.path)
-                        .put("directory", file.directory)
-                        .put("size", file.size)
-                        .put("modified", file.modified),
-                )
-            }
-        })
-
-    private fun resolveEmailAccount(args: JSONObject): EmailServerConfig =
-        settings.resolveEmailServer(args.getString("account_id"))
-            ?: error("Email account is missing or disabled: ${args.optString("account_id")}. Call list_email_accounts and use a returned id.")
-
-    private fun emailComposeRequest(args: JSONObject): EmailComposeRequest {
-        fun strings(name: String): List<String> = args.optJSONArray(name)?.let { array ->
-            buildList {
-                for (index in 0 until array.length()) {
-                    array.optString(index).trim().takeIf { it.isNotBlank() }?.let(::add)
-                }
-            }
-        }.orEmpty()
-        val attachmentPaths = strings("attachments")
-        val attachments = attachmentPaths.map { path ->
-            val bytes = nativeFileManager.readBytes(path, EmailClient.MAX_ATTACHMENT_BYTES.toLong() + 1L).getOrThrow()
-            require(bytes.size <= EmailClient.MAX_ATTACHMENT_BYTES) {
-                "Attachment exceeds 20 MB: $path. Use a cloud link or file-transfer service instead."
-            }
-            OutgoingAttachment(path.substringAfterLast('/').substringAfterLast('\\'), bytes)
-        }
-        require(attachments.sumOf { it.bytes.size.toLong() } <= EmailClient.MAX_ATTACHMENT_BYTES) {
-            "Combined attachments exceed 20 MB. Use a cloud link or file-transfer service instead."
-        }
-        return EmailComposeRequest(
-            to = strings("to"),
-            cc = strings("cc"),
-            bcc = strings("bcc"),
-            subject = args.getString("subject"),
-            textBody = args.optString("text_body"),
-            htmlBody = args.optString("html_body"),
-            attachments = attachments,
-            replyFolder = args.optString("reply_folder"),
-            replyUid = args.optLong("reply_uid", 0L),
-            allowReplyToAnswered = args.optBoolean("allow_reply_to_answered", false),
-        )
-    }
-
-    private fun parseBackupOptions(args: JSONObject): BackupOptions = BackupOptions(
-        includeProfile = args.optBoolean("include_profile", true),
-        includeConversations = args.optBoolean("include_conversations", true),
-        includeModelProfiles = args.optBoolean("include_model_profiles", true),
-        includeMcp = args.optBoolean("include_mcp", true),
-        includeSsh = args.optBoolean("include_ssh", true),
-        includeEmail = args.optBoolean("include_email", true),
-        includePrompts = args.optBoolean("include_prompts", true),
-        includeMemories = args.optBoolean("include_memories", true),
-        includeSkills = args.optBoolean("include_skills", true),
-        includeWebDav = args.optBoolean("include_webdav", true),
-        includeFileTransfer = args.optBoolean("include_file_transfer", true),
-        includeSecrets = args.optBoolean("include_secrets", false),
-    )
-
-    private fun skillsJson(): JSONArray = JSONArray().also { array ->
-        settings.installedSkills().forEach { array.put(skillJson(it)) }
-    }
-
-    private fun skillJson(skill: SkillPack): JSONObject = JSONObject()
-        .put("id", skill.id)
-        .put("name", skill.name)
-        .put("description", skill.description)
-        .put("enabled", skill.enabled)
-        .put("file_count", skill.fileCount)
-
-    private fun agentToolsJson(): JSONArray {
-        val disabled = settings.disabledTools()
-        val mcpToolMeta = allMcpToolMetaForConfig()
-        val names = agentToolNamesForConfig()
-        return JSONArray().also { array ->
-            names.forEach { name ->
-                val mcpMeta = mcpToolMeta[name]
-                val serverEnabled = mcpMeta?.first ?: true
-                val item = JSONObject()
-                    .put("name", name)
-                    .put("enabled", name == "manage_app_config" || (name !in disabled && serverEnabled))
-                    .put("deletable", false)
-                    .put("protected", name == "manage_app_config")
-                item.apply {
-                    mcpMeta?.let { (mcpServerEnabled, serverName, toolName) ->
-                        put("source", "mcp")
-                        put("server_enabled", mcpServerEnabled)
-                        put("server_name", serverName)
-                        put("mcp_tool", toolName)
-                        put("tool_enabled", name !in disabled)
-                        put("available_in_prompt", mcpServerEnabled && name !in disabled)
-                    } ?: put("source", "local")
-                }
-                array.put(item)
-            }
-        }
-    }
-
-    private fun agentToolNamesForConfig(): List<String> {
-        return (CONFIGURABLE_AGENT_TOOLS + allMcpToolMetaForConfig().keys)
-            .distinct()
-            .sorted()
-    }
-
-    private fun manageScheduledTasks(args: JSONObject): String {
-        val action = args.optString("action").trim().lowercase(Locale.US)
-        if (action == "list") {
-            return JSONObject()
-                .put("schema", "lyra_scheduled_tasks_v1")
-                .put("tasks", scheduledTaskManager.describe())
-                .toString()
-        }
-        val taskId = args.optString("task_id")
-        if (action == "delete") {
-                require(taskId.isNotBlank()) { "task_id is required." }
-            scheduledTaskManager.delete(taskId)
-            return JSONObject().put("ok", true).put("action", action).put("task_id", taskId).toString()
-        }
-        if (action == "enable" || action == "disable") {
-                require(taskId.isNotBlank()) { "task_id is required." }
-            val task = scheduledTaskManager.setEnabled(taskId, action == "enable")
-                    ?: error("Scheduled task does not exist: $taskId. Call action=list and use a returned task_id.")
-            return scheduledTaskResult(action, task)
-        }
-        require(action == "create" || action == "update") { "action must be list, create, update, enable, disable, or delete." }
-        val existing = taskId.takeIf { it.isNotBlank() }?.let(scheduledTaskManager::task)
-        if (action == "update") require(existing != null) { "Scheduled task does not exist: $taskId. Call action=list and use a returned task_id." }
-        val profile = settings.profiles().firstOrNull { it.id == args.optString("profile_id") }
-            ?: existing?.profileId?.let { id -> settings.profiles().firstOrNull { it.id == id } }
-            ?: settings.selectedProfile()
-        val type = args.optString("schedule_type", existing?.type?.name.orEmpty())
-            .uppercase(Locale.US)
-            .let { runCatching { ScheduledTaskType.valueOf(it) }.getOrDefault(existing?.type ?: ScheduledTaskType.ONCE) }
-        val runAt = args.optString("run_at").takeIf { it.isNotBlank() }?.let(::parseAgentTime)
-            ?: existing?.runAtMillis
-            ?: 0L
-        if (type == ScheduledTaskType.ONCE) require(runAt > System.currentTimeMillis()) {
-            "run_at for a one-time task must be in the future and use yyyy-MM-dd HH:mm or ISO-8601."
-        }
-        val task = ScheduledTask(
-            id = existing?.id ?: java.util.UUID.randomUUID().toString(),
-            title = args.optString("title").ifBlank { existing?.title ?: "定时任务" },
-            prompt = args.optString("prompt").ifBlank { existing?.prompt.orEmpty() },
-            type = type,
-            hour = if (args.has("hour")) args.optInt("hour") else existing?.hour ?: 9,
-            minute = if (args.has("minute")) args.optInt("minute") else existing?.minute ?: 0,
-            runAtMillis = runAt,
-            dayOfWeek = if (args.has("day_of_week")) args.optInt("day_of_week") else existing?.dayOfWeek ?: 1,
-            dayOfMonth = if (args.has("day_of_month")) args.optInt("day_of_month") else existing?.dayOfMonth ?: 1,
-            profileId = profile.id,
-            model = args.optString("model").ifBlank { existing?.model ?: profile.selectedModel },
-            enabled = if (args.has("enabled")) args.optBoolean("enabled") else existing?.enabled ?: true,
-            createdAt = existing?.createdAt ?: System.currentTimeMillis(),
-            lastRunAt = existing?.lastRunAt ?: 0L,
-            finishedAt = existing?.finishedAt ?: 0L,
-            status = existing?.status ?: com.yukisoffd.lyracode.tasks.ScheduledTaskStatus.IDLE,
-            result = existing?.result.orEmpty(),
-            error = existing?.error.orEmpty(),
-        )
-        require(task.prompt.isNotBlank()) { "prompt is required." }
-        return scheduledTaskResult(action, scheduledTaskManager.save(task))
-    }
-
-    private fun scheduledTaskResult(action: String, task: ScheduledTask): String = JSONObject()
-        .put("ok", true)
-        .put("action", action)
-        .put("task_id", task.id)
-        .put("title", task.title)
-        .put("schedule_type", task.type.name.lowercase(Locale.US))
-        .put("enabled", task.enabled)
-        .put("next_run_at", task.nextRunAt)
-        .put("profile_id", task.profileId)
-        .put("model", task.model)
-        .toString()
-
-    private fun manageMiniServer(args: JSONObject): String {
-        val action = args.optString("action", "status").lowercase(Locale.US)
-        val current = settings.miniServerConfig()
-        val config = current.copy(
-            protocol = args.optString("protocol").ifBlank { current.protocol }.lowercase(Locale.US).let {
-                if (it == AppSettings.MINI_SERVER_PROTOCOL_HTTPS) AppSettings.MINI_SERVER_PROTOCOL_HTTPS else AppSettings.MINI_SERVER_PROTOCOL_HTTP
-            },
-            host = args.optString("host").ifBlank { current.host },
-            port = if (args.has("port")) args.optInt("port", current.port).coerceIn(1, 65535) else current.port,
-            username = args.optString("username").ifBlank { current.username },
-            password = if (args.has("password")) args.optString("password") else current.password,
-            customDomains = miniServerDomains(args, current.customDomains),
-            forceHttps = if (args.has("force_https")) args.optBoolean("force_https") else current.forceHttps,
-            tlsKeyStoreBase64 = if (args.has("tls_key_store_base64")) args.optString("tls_key_store_base64") else current.tlsKeyStoreBase64,
-            tlsKeyStorePassword = if (args.has("tls_key_store_password")) args.optString("tls_key_store_password") else current.tlsKeyStorePassword,
-            tlsCertificateChain = if (args.has("tls_certificate_chain")) args.optString("tls_certificate_chain") else current.tlsCertificateChain,
-            tlsPrivateKey = if (args.has("tls_private_key")) args.optString("tls_private_key") else current.tlsPrivateKey,
-            spaFallback = if (args.has("spa_fallback")) args.optBoolean("spa_fallback") else current.spaFallback,
-            directoryListing = if (args.has("directory_listing")) args.optBoolean("directory_listing") else current.directoryListing,
-            mdnsEnabled = if (args.has("mdns_enabled")) args.optBoolean("mdns_enabled") else current.mdnsEnabled,
-            mdnsName = args.optString("mdns_name").ifBlank { current.mdnsName },
-        )
-        val status = when (action) {
-            "status" -> miniServerManager.status()
-            "update" -> {
-                settings.saveMiniServerConfig(config)
-                miniServerManager.status()
-            }
-            "start" -> miniServerManager.start(config.copy(enabled = true))
-            "stop" -> miniServerManager.stop()
-            "restart" -> miniServerManager.restart(config.copy(enabled = true))
-            "reset" -> {
-                if (miniServerManager.status().running) miniServerManager.stop()
-                settings.saveMiniServerConfig(
-                    MiniServerConfig(
-                        protocol = AppSettings.MINI_SERVER_PROTOCOL_HTTP,
-                        host = AppSettings.DEFAULT_MINI_SERVER_HOST,
-                        port = AppSettings.DEFAULT_MINI_SERVER_PORT,
-                        username = AppSettings.DEFAULT_MINI_SERVER_USERNAME,
-                        password = "",
-                        customDomains = emptyList(),
-                        forceHttps = false,
-                        tlsKeyStoreBase64 = "",
-                        tlsKeyStorePassword = "",
-                        tlsCertificateChain = "",
-                        tlsPrivateKey = "",
-                        spaFallback = true,
-                        directoryListing = false,
-                        mdnsEnabled = false,
-                        mdnsName = AppSettings.DEFAULT_MINI_SERVER_MDNS_NAME,
-                        enabled = false,
-                    ),
-                )
-                miniServerManager.status()
-            }
-            else -> error("Unknown mini-server action: $action. Use status, update, start, stop, restart, or reset.")
-        }
-        return miniServerManager.statusJson()
-            .put("action", action)
-            .put("running", status.running)
-            .put("security_note", miniServerSecurityNote(config))
-            .toString()
-    }
-
-    private fun readMiniServerLogs(args: JSONObject): String {
-        val limit = args.optInt("limit", 120).coerceIn(1, 500)
-        val level = args.optString("level").lowercase(Locale.US).takeIf { it in setOf("debug", "info", "warn", "error") }.orEmpty()
-        return miniServerManager.logsJson(limit, level).toString()
-    }
-
-    private fun miniServerDomains(args: JSONObject, current: List<String>): List<String> {
-        val array = args.optJSONArray("custom_domains")
-        if (array != null) {
-            return buildList {
-                for (index in 0 until array.length()) {
-                    array.optString(index).trim().takeIf { it.isNotBlank() }?.let(::add)
-                }
-            }.distinct()
-        }
-        return args.optString("custom_domains")
-            .takeIf { it.isNotBlank() }
-            ?.lineSequence()
-            ?.map { it.trim() }
-            ?.filter { it.isNotBlank() }
-            ?.distinct()
-            ?.toList()
-            ?: current
-    }
-
-    private fun miniServerSecurityNote(config: MiniServerConfig): String {
-        return buildString {
-            append("The mini server uses the current workspace as its static site root.")
-            if (config.protocol == AppSettings.MINI_SERVER_PROTOCOL_HTTPS) {
-                append(" HTTPS uses the built-in self-signed certificate, which browsers may distrust. Use trusted TLS through a reverse proxy or tunnel for public sharing.")
-            }
-            if (config.forceHttps) {
-                append(" Forced HTTPS is enabled; HTTP requests are redirected.")
-            }
-            if (config.customDomains.isNotEmpty()) {
-                append(" Configured domains: ${config.customDomains.joinToString(", ")}.")
-            }
-            if (config.host == "0.0.0.0" || config.host == "::") {
-                append(" The bind address exposes the server to the LAN and potentially the public internet through port mapping or tunneling.")
-            }
-            if (config.password.isBlank()) {
-                    append(" No access password is configured; use only on a trusted network.")
-            }
-            if (config.protocol == AppSettings.MINI_SERVER_PROTOCOL_HTTP) {
-                append(" Plain HTTP can expose paths, content, and credentials.")
-            }
-        }
-    }
-
-    private fun readMemories(args: JSONObject): String {
-        val query = args.optString("query").trim()
-        val includeDisabled = args.optBoolean("include_disabled", false)
-        val entries = settings.memories().filter { memory ->
-            (includeDisabled || memory.enabled) &&
-                (query.isBlank() ||
-                    memory.content.contains(query, ignoreCase = true) ||
-                    memory.category.contains(query, ignoreCase = true))
-        }
-        return JSONObject()
-            .put("schema", "lyra_user_memories_v1")
-            .put("count", entries.size)
-            .put("memories", JSONArray().also { array -> entries.forEach { array.put(memoryJson(it)) } })
-            .toString()
-    }
-
-    private fun saveMemory(args: JSONObject): String {
-        val memory = settings.createMemory(
-            content = args.getString("content"),
-            category = args.optString("category", MemoryEntry.CATEGORY_OTHER),
-        )
-        return JSONObject()
-            .put("schema", "lyra_user_memory_change_v1")
-            .put("action", "saved")
-            .put("memory", memoryJson(memory))
-            .toString()
-    }
-
-    private fun updateMemory(args: JSONObject): String {
-        require(args.has("content") || args.has("category") || args.has("enabled")) {
-            "update_memory requires at least one of content, category, or enabled."
-        }
-        val memory = settings.updateMemory(
-            id = args.getString("id"),
-            content = args.optString("content").takeIf { args.has("content") },
-            category = args.optString("category").takeIf { args.has("category") },
-            enabled = args.optBoolean("enabled").takeIf { args.has("enabled") },
-        )
-        return JSONObject()
-            .put("schema", "lyra_user_memory_change_v1")
-            .put("action", "updated")
-            .put("memory", memoryJson(memory))
-            .toString()
-    }
-
-    private fun deleteMemory(args: JSONObject): String {
-        val id = args.getString("id")
-        require(settings.deleteMemory(id)) { "Memory does not exist: $id. Call read_memories and use a returned id." }
-        return JSONObject()
-            .put("schema", "lyra_user_memory_change_v1")
-            .put("action", "deleted")
-            .put("id", id)
-            .toString()
-    }
-
-    private fun memoryJson(memory: MemoryEntry): JSONObject = JSONObject()
-        .put("id", memory.id)
-        .put("content", memory.content)
-        .put("category", memory.category)
-        .put("enabled", memory.enabled)
-        .put("created_at", memory.createdAt)
-        .put("updated_at", memory.updatedAt)
-
-    private fun searchConversationHistory(args: JSONObject): String {
-        val query = args.optString("query").trim()
-        val start = args.optString("start_time").takeIf { it.isNotBlank() }?.let(::parseAgentTime) ?: Long.MIN_VALUE
-        val end = args.optString("end_time").takeIf { it.isNotBlank() }?.let(::parseAgentTime) ?: Long.MAX_VALUE
-        val limit = args.optInt("limit", 20).coerceIn(1, 100)
-        val results = conversationStore.conversations(ConversationStore.MODE_NORMAL)
-            .asSequence()
-            .filter { it.updatedAt in start..end }
-            .map { conversation ->
-                val visible = conversationStore.messages(conversation.id)
-                    .filter { it.role == "user" || it.role == "assistant" }
-                    .filter { it.content.isNotBlank() }
-                conversation to visible
-            }
-            .filter { (conversation, messages) ->
-                query.isBlank() ||
-                    conversation.title.contains(query, ignoreCase = true) ||
-                    messages.any { it.content.contains(query, ignoreCase = true) }
-            }
-            .take(limit)
-            .toList()
-        return JSONObject()
-            .put("schema", "lyra_conversation_search_v1")
-            .put("thinking_included", false)
-            .put("tool_messages_included", false)
-            .put(
-                "conversations",
-                JSONArray().also { array ->
-                    results.forEach { (conversation, messages) ->
-                        array.put(
-                            JSONObject()
-                                .put("id", conversation.id.toString())
-                                .put("title", conversation.title)
-                                .put("created_at", conversation.createdAt)
-                                .put("updated_at", conversation.updatedAt)
-                                .put("message_count", messages.size)
-                                .put(
-                                    "preview",
-                                    messages.asReversed().firstOrNull()?.content.orEmpty()
-                                        .replace(Regex("\\s+"), " ")
-                                        .take(240),
-                                ),
-                        )
-                    }
-                },
-            )
-            .toString()
-    }
-
-    private fun readConversationHistory(args: JSONObject): String {
-        val ids = buildList {
-            args.optJSONArray("conversation_ids")?.let { array ->
-                for (index in 0 until array.length()) {
-                    array.optString(index).toLongOrNull()?.let(::add)
-                }
-            }
-            args.optString("conversation_id").toLongOrNull()?.let(::add)
-        }.distinct().take(20)
-        require(ids.isNotEmpty()) { "Provide conversation_id or a non-empty conversation_ids array." }
-        val maxMessages = args.optInt("max_messages", 100).coerceIn(1, 500)
-        val conversations = JSONArray()
-        ids.forEach { id ->
-            val conversation = conversationStore.conversation(id)
-            if (conversation == null || conversation.mode != ConversationStore.MODE_NORMAL) return@forEach
-            val visible = conversationStore.messages(id)
-                .filter { it.role == "user" || it.role == "assistant" }
-                .filter { it.content.isNotBlank() }
-                .takeLast(maxMessages)
-            conversations.put(
-                JSONObject()
-                    .put("id", id.toString())
-                    .put("title", conversation.title)
-                    .put("created_at", conversation.createdAt)
-                    .put("updated_at", conversation.updatedAt)
-                    .put(
-                        "messages",
-                        JSONArray().also { array ->
-                            visible.forEach { message ->
-                                array.put(
-                                    JSONObject()
-                                        .put("role", message.role)
-                                        .put("content", message.content)
-                                        .put("created_at", message.createdAt),
-                                )
-                            }
-                        },
-                    ),
-            )
-        }
-        return JSONObject()
-            .put("schema", "lyra_conversation_history_v1")
-            .put("thinking_included", false)
-            .put("tool_messages_included", false)
-            .put("conversations", conversations)
-            .toString()
-    }
-
-    private fun parseAgentTime(value: String): Long {
-        value.toLongOrNull()?.let { return it }
-        val zone = ZoneId.systemDefault()
-        val dateTimeFormats = listOf(
-            DateTimeFormatter.ISO_LOCAL_DATE_TIME,
-            DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"),
-            DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm"),
-        )
-        dateTimeFormats.forEach { formatter ->
-            try {
-                return LocalDateTime.parse(value.trim(), formatter).atZone(zone).toInstant().toEpochMilli()
-            } catch (_: DateTimeParseException) {
-            }
-        }
-        return try {
-            LocalDate.parse(value.trim(), DateTimeFormatter.ISO_LOCAL_DATE)
-                .atStartOfDay(zone)
-                .toInstant()
-                .toEpochMilli()
-        } catch (_: DateTimeParseException) {
-            error("Cannot parse time: $value. Use an epoch timestamp, yyyy-MM-dd, yyyy-MM-dd HH:mm, or ISO-8601.")
-        }
-    }
-
-    private fun allMcpToolMetaForConfig(): Map<String, Triple<Boolean, String, String>> {
-        return buildMap {
-            settings.mcpServers().forEach { server ->
-                server.tools.forEach { tool ->
-                    put(settings.mcpToolFunctionName(server, tool), Triple(server.enabled, server.name, tool.name))
-                }
-            }
-        }
     }
 
     private fun approvalFor(conversationId: Long, call: ToolCall): ToolApprovalRequest? {
@@ -4615,7 +3062,7 @@ class OpenAiAgent(
             } else {
                 null
             }
-            "run_command" -> if (requiresCommandApproval(args.optString("command"))) {
+            "run_command" -> if (fileTools.requiresCommandApproval(args.optString("command"))) {
                 ToolApprovalRequest(
                     conversationId,
                     call.name,
@@ -4626,7 +3073,7 @@ class OpenAiAgent(
             } else {
                 null
             }
-            "proot_command" -> if (requiresCommandApproval(args.toolCommandArgument())) {
+            "proot_command" -> if (fileTools.requiresCommandApproval(args.toolCommandArgument())) {
                 ToolApprovalRequest(
                     conversationId,
                     call.name,
@@ -4773,243 +3220,6 @@ class OpenAiAgent(
                 null
             }
         }
-    }
-
-    private suspend fun writeFileWithDiff(path: String, content: String): ToolExecution {
-        val before = nativeFileManager.readFileForEdit(path).getOrNull().orEmpty()
-        return withFileActivity(path, globalStorage = false, operation = "write", content = before) {
-            val editorApplied = applyFileChangeInEditor(path, before, content, globalStorage = false)
-            val message = nativeFileManager.writeFile(path, content).getOrThrow()
-            val after = nativeFileManager.readFileForEdit(path).getOrNull().orEmpty()
-            fileMutationHandler(
-                AgentFileMutation(path, after, globalStorage = false, beforeContent = before, editorApplied = editorApplied),
-            )
-            appendDiff(message, path, before, after)
-        }
-    }
-
-    private suspend fun appendFileWithDiff(path: String, content: String): ToolExecution {
-        val before = nativeFileManager.readFileForEdit(path).getOrNull().orEmpty()
-        return withFileActivity(path, globalStorage = false, operation = "append", content = before) {
-            val expectedAfter = before + content
-            val editorApplied = applyFileChangeInEditor(path, before, expectedAfter, globalStorage = false)
-            val message = nativeFileManager.appendFile(path, content).getOrThrow()
-            val after = nativeFileManager.readFileForEdit(path).getOrNull().orEmpty()
-            fileMutationHandler(
-                AgentFileMutation(path, after, globalStorage = false, beforeContent = before, editorApplied = editorApplied),
-            )
-            appendDiff(message, path, before, after)
-        }
-    }
-
-    private suspend fun globalWriteFileWithDiff(path: String, content: String): ToolExecution {
-        val before = globalFileManager.readFileForEdit(path).getOrNull().orEmpty()
-        return withFileActivity(path, globalStorage = true, operation = "write", content = before) {
-            val editorApplied = applyFileChangeInEditor(path, before, content, globalStorage = true)
-            val message = globalFileManager.writeFile(path, content).getOrThrow()
-            val after = globalFileManager.readFileForEdit(path).getOrNull().orEmpty()
-            fileMutationHandler(
-                AgentFileMutation(path, after, globalStorage = true, beforeContent = before, editorApplied = editorApplied),
-            )
-            appendDiff(message, path, before, after)
-        }
-    }
-
-    private suspend fun globalAppendFileWithDiff(path: String, content: String): ToolExecution {
-        val before = globalFileManager.readFileForEdit(path).getOrNull().orEmpty()
-        return withFileActivity(path, globalStorage = true, operation = "append", content = before) {
-            val expectedAfter = before + content
-            val editorApplied = applyFileChangeInEditor(path, before, expectedAfter, globalStorage = true)
-            val message = globalFileManager.appendFile(path, content).getOrThrow()
-            val after = globalFileManager.readFileForEdit(path).getOrNull().orEmpty()
-            fileMutationHandler(
-                AgentFileMutation(path, after, globalStorage = true, beforeContent = before, editorApplied = editorApplied),
-            )
-            appendDiff(message, path, before, after)
-        }
-    }
-
-    private suspend fun editFileWithDiff(args: JSONObject, globalStorage: Boolean): ToolExecution {
-        val path = args.getString("path")
-        val before = if (globalStorage) {
-            globalFileManager.readFileForEdit(path).getOrThrow()
-        } else {
-            nativeFileManager.readFileForEdit(path).getOrThrow()
-        }
-        return withFileActivity(path, globalStorage, operation = "edit", content = before) {
-            val usesLineRange = args.has("start_line") || args.has("end_line")
-            val usesExactMatch = args.has("old_content") || args.has("old_content_lines")
-            require(usesLineRange.xor(usesExactMatch)) {
-                "Choose exactly one edit mode: start_line/end_line or old_content/old_content_lines."
-            }
-            val newContent = args.toolTextArgument("new_content")
-            val after = if (usesLineRange) {
-                val startLine = args.getInt("start_line")
-                applyLineRangeReplacement(
-                    source = before,
-                    startLine = startLine,
-                    endLine = args.optInt("end_line", startLine),
-                    newContent = newContent,
-                )
-            } else {
-                applyExactTextReplacement(
-                    source = before,
-                    oldContent = args.toolTextArgument("old_content"),
-                    newContent = newContent,
-                    expectedReplacements = args.optInt("expected_replacements", 1),
-                )
-            }
-            require(after != before) { "The edit would not change the file, so no write was performed. Re-read the target context and correct the edit." }
-            val editorApplied = applyFileChangeInEditor(path, before, after, globalStorage)
-            val message = if (globalStorage) {
-                globalFileManager.writeFile(path, after).getOrThrow()
-            } else {
-                nativeFileManager.writeFile(path, after).getOrThrow()
-            }
-            fileMutationHandler(
-                AgentFileMutation(path, after, globalStorage, beforeContent = before, editorApplied = editorApplied),
-            )
-            appendDiff(message, path, before, after)
-        }
-    }
-
-    private suspend fun applyFileChangeInEditor(
-        path: String,
-        before: String,
-        after: String,
-        globalStorage: Boolean,
-    ): Boolean {
-        val result = fileEditHandler(
-            AgentFileMutation(
-                path = path,
-                content = after,
-                globalStorage = globalStorage,
-                beforeContent = before,
-            ),
-        )
-        if (result.handled && !result.applied) {
-            error(result.message.ifBlank { "The file editor could not apply the change; the disk write was cancelled. Re-read the current file and retry with exact context." })
-        }
-        return result.applied
-    }
-
-    private suspend fun readFileWithActivity(path: String, globalStorage: Boolean): ToolExecution {
-        val content = if (globalStorage) {
-            globalFileManager.readFile(path).getOrThrow()
-        } else {
-            nativeFileManager.readFile(path).getOrThrow()
-        }
-        return withFileActivity(path, globalStorage, operation = "read", content = content) {
-            ToolExecution(content)
-        }
-    }
-
-    private suspend fun readFileLines(args: JSONObject, globalStorage: Boolean): String {
-        val path = args.getString("path")
-        val startLine = args.optInt("start_line", 1).coerceAtLeast(1)
-        val lineCount = args.optInt("line_count", 200).coerceIn(1, 1_000)
-        val content = if (globalStorage) {
-            globalFileManager.readFileForEdit(path).getOrThrow()
-        } else {
-            nativeFileManager.readFileForEdit(path).getOrThrow()
-        }
-        val lines = content.replace("\r\n", "\n").replace('\r', '\n').split('\n')
-        return withFileActivity(path, globalStorage, operation = "read", content = content) {
-            if (startLine > lines.size) {
-                return@withFileActivity "FILE_LINES path=$path total_lines=${lines.size}\nRequested start_line $startLine is outside the file. Retry with a line number from 1 to ${lines.size}."
-            }
-            val endExclusive = (startLine - 1 + lineCount).coerceAtMost(lines.size)
-            val body = buildString {
-                for (index in startLine - 1 until endExclusive) {
-                    append(index + 1).append("| ").append(lines[index]).append('\n')
-                    if (length >= 240_000) {
-                        append("...output reached the 240000-character limit; retry with a smaller line_count.\n")
-                        break
-                    }
-                }
-            }
-            "FILE_LINES path=$path range=$startLine-$endExclusive total_lines=${lines.size}\n$body"
-        }
-    }
-
-    private suspend fun <T> withFileActivity(
-        path: String,
-        globalStorage: Boolean,
-        operation: String,
-        content: String?,
-        block: suspend () -> T,
-    ): T {
-        fileActivityHandler(AgentFileActivity(path, globalStorage, operation, content))
-        return try {
-            delay(90L)
-            block()
-        } finally {
-            fileActivityHandler(null)
-        }
-    }
-
-    private fun deleteWithDiff(path: String): ToolExecution {
-        val before = nativeFileManager.readFile(path).getOrNull().orEmpty()
-        val message = nativeFileManager.delete(path).getOrThrow()
-        return appendDiff(message, path, before, "")
-    }
-
-    private fun renameMoveWithDiff(from: String, to: String): ToolExecution {
-        val before = nativeFileManager.readFile(from).getOrNull().orEmpty()
-        val message = nativeFileManager.renameMove(from, to).getOrThrow()
-        val after = nativeFileManager.readFile(to).getOrNull().orEmpty()
-        return appendDiff(message, to, before, after)
-    }
-
-    private fun appendDiff(message: String, path: String, before: String, after: String): ToolExecution {
-        val diff = FileDiff.from(path, before, after)
-        return ToolExecution(message, listOf(diff))
-    }
-
-    private fun requiresCommandApproval(command: String): Boolean {
-        val lowered = command.lowercase()
-        val readOnlyCommands = listOf("pwd", "ls", "cat", "head", "tail", "grep", "find", "awk")
-        val first = lowered.trim().split(Regex("\\s+")).firstOrNull().orEmpty().substringAfterLast("/")
-        if (first !in readOnlyCommands) return true
-        val mutatingFragments = listOf(
-            ">", ">>", "| tee", " rm ", " mv ", " cp ", " mkdir ", " touch ", " chmod ", " sed -i",
-            "pip install", "npm install", "pnpm install", "yarn add", "apt ", "pkg ", "git ",
-            "python ", "python3 ", "node ",
-        )
-        val padded = " $lowered "
-        return mutatingFragments.any { padded.contains(it) }
-    }
-
-    private fun isFileSearchCommand(command: String): Boolean {
-        val lowered = command.lowercase()
-        return FILE_SEARCH_COMMAND_PATTERNS.any { it.containsMatchIn(lowered) }
-    }
-
-    private fun globalSearchFiles(query: String): ToolExecution {
-        val cleanQuery = query.trim()
-        require(cleanQuery.isNotBlank()) { "Search query must not be empty." }
-        val result = globalFileManager.searchFiles(cleanQuery, GLOBAL_SEARCH_RESULT_LIMIT).getOrThrow()
-        return ToolExecution(
-            "GLOBAL_SEARCH_FILES_RESULT\n" +
-                "root=/storage/emulated/0\n" +
-                "query=$cleanQuery\n" +
-                "limit=$GLOBAL_SEARCH_RESULT_LIMIT\n" +
-                "note=These results are outside the workspace and use absolute shared-storage paths. Read them with global_read_file/global_read_file_lines and modify them only with matching global_* tools.\n" +
-                result.toAgentText(),
-        )
-    }
-
-    private fun JSONObject.toolCommandArgument(): String {
-        val lines = optJSONArray("command_lines")
-        if (lines != null && lines.length() > 0) {
-            return buildString {
-                for (index in 0 until lines.length()) {
-                    if (index > 0) append('\n')
-                    append(lines.optString(index))
-                }
-            }
-        }
-        return stringFieldOrNull("command") ?: error("A command tool requires command or command_lines.")
     }
 
     private fun titleFor(conversationId: Long, userInput: String): String? {
@@ -5539,26 +3749,6 @@ class OpenAiAgent(
         return conversationStore.conversation(conversationId)?.mode == ConversationStore.MODE_SUBAGENT
     }
 
-    private fun List<WorkspaceFile>.toAgentText(): String {
-        if (isEmpty()) return "(empty)"
-        return joinToString("\n") {
-            val type = if (it.directory) "dir " else "file"
-            "$type\t${it.size}\t${it.path}"
-        }
-    }
-
-    private fun List<WorkspaceFile>.toSearchAgentText(query: String, path: String): String {
-        val cleanPath = path.trim().ifBlank { "." }
-        if (isEmpty()) {
-            return "SEARCH_EMPTY\n" +
-                "query=$query\n" +
-                "path=$cleanPath\n" +
-                "workspace=${workspaceManager.displayName()}\n" +
-                "note=Only the authorized workspace was searched. If the target may be outside it, use global_search_files for Android shared storage."
-        }
-        return toAgentText()
-    }
-
     private fun splitInlineThink(content: String, existingThinking: String): Pair<String, String> {
         val start = content.indexOf("<think>", ignoreCase = true)
         val end = content.indexOf("</think>", ignoreCase = true)
@@ -5636,10 +3826,8 @@ class OpenAiAgent(
         private const val HISTORY_COMPRESSION_FINAL_MAX_OUTPUT_TOKENS = 4096
         private const val PROMPT_CACHE_KEY_HASH_CHARS = 32
         private const val MESSAGE_WRAPPER_TOKENS = 8L
-        private const val GLOBAL_SEARCH_RESULT_LIMIT = 120
         private const val MAX_IMAGE_PROMPT_BYTES = 8 * 1024 * 1024
         private const val MAX_VISION_SUPPLEMENT_IMAGES = 8
-        private const val DEFAULT_WEBDAV_BACKUP_PATH = "/LyraCode/lyra_backup_latest.zip"
         private const val LOCAL_MCP_CONVERSATION_ID = 0L
         private val JSON_SCHEMA_TYPES = setOf("string", "number", "integer", "boolean", "object", "array")
         private val FILE_TEXT_ARGUMENT_TOOLS = setOf(
@@ -5777,139 +3965,7 @@ class OpenAiAgent(
             "update_todo_item",
             "proot_command",
         )
-        private val FILE_SEARCH_COMMAND_PATTERNS = listOf(
-            Regex("""(^|[;&|()\n]\s*)find\s+.+\s-(i)?name\s+"""),
-            Regex("""(^|[;&|()\n]\s*)fd\s+"""),
-            Regex("""(^|[;&|()\n]\s*)fdfind\s+"""),
-            Regex("""(^|[;&|()\n]\s*)locate\s+"""),
-        )
     }
-}
-
-private fun ToolExecution.toToolOutputJson(toolName: String, ok: Boolean): String {
-    return JSONObject()
-        .put("schema", "lyra_tool_output_v2")
-        .put("ok", ok)
-        .put("tool", toolName)
-        .put("content", content)
-        .put("error", if (ok) "" else content)
-        .put("file_changes", JSONArray().apply { fileChanges.forEach { put(it.toJson()) } })
-        .toString()
-}
-
-private data class FileDiff(
-    val path: String,
-    val added: Int,
-    val removed: Int,
-    val diff: String,
-    val before: String,
-    val after: String,
-) {
-    fun toJson(): JSONObject {
-        return JSONObject()
-            .put("path", path)
-            .put("added", added)
-            .put("removed", removed)
-            .put("diff", diff)
-            .put("before", before)
-            .put("after", after)
-    }
-
-    fun toToolText(): String {
-        return """
-        LYRA_FILE_CHANGE_BEGIN
-        path: $path
-        added: $added
-        removed: $removed
-        diff:
-        $diff
-        LYRA_FILE_BEFORE_BEGIN
-        $before
-        LYRA_FILE_BEFORE_END
-        LYRA_FILE_AFTER_BEGIN
-        $after
-        LYRA_FILE_AFTER_END
-        LYRA_FILE_CHANGE_END
-        """.trimIndent()
-    }
-
-    companion object {
-        fun from(path: String, before: String, after: String): FileDiff {
-            val beforeLines = before.toDiffLines()
-            val afterLines = after.toDiffLines()
-            val lcs = Array(beforeLines.size + 1) { IntArray(afterLines.size + 1) }
-            for (i in beforeLines.indices.reversed()) {
-                for (j in afterLines.indices.reversed()) {
-                    lcs[i][j] = if (beforeLines[i] == afterLines[j]) {
-                        lcs[i + 1][j + 1] + 1
-                    } else {
-                        maxOf(lcs[i + 1][j], lcs[i][j + 1])
-                    }
-                }
-            }
-            val diffLines = mutableListOf<String>()
-            var added = 0
-            var removed = 0
-            var i = 0
-            var j = 0
-            while (i < beforeLines.size && j < afterLines.size) {
-                when {
-                    beforeLines[i] == afterLines[j] -> {
-                        diffLines += "  ${beforeLines[i]}"
-                        i++
-                        j++
-                    }
-                    lcs[i + 1][j] >= lcs[i][j + 1] -> {
-                        diffLines += "- ${beforeLines[i]}"
-                        removed++
-                        i++
-                    }
-                    else -> {
-                        diffLines += "+ ${afterLines[j]}"
-                        added++
-                        j++
-                    }
-                }
-            }
-            while (i < beforeLines.size) {
-                diffLines += "- ${beforeLines[i++]}"
-                removed++
-            }
-            while (j < afterLines.size) {
-                diffLines += "+ ${afterLines[j++]}"
-                added++
-            }
-            return FileDiff(
-                path = path,
-                added = added,
-                removed = removed,
-                diff = diffLines.take(2_000).joinToString("\n"),
-                before = before.take(20_000),
-                after = after.take(20_000),
-            )
-        }
-
-        private fun String.toDiffLines(): List<String> {
-            if (isEmpty()) return emptyList()
-            return replace("\r\n", "\n").lines()
-        }
-    }
-}
-
-private fun JSONObject.cleanString(name: String): String {
-    return stringFieldOrNull(name).orEmpty()
-}
-
-private fun JSONObject.stringFieldOrNull(name: String): String? {
-    if (!has(name) || isNull(name)) return null
-    val value = opt(name) ?: return null
-    val text = value as? String ?: return null
-    return text.takeUnless { it.equals("null", ignoreCase = true) }
-}
-
-private fun JSONObject.booleanOrNull(name: String): Boolean? {
-    if (!has(name) || isNull(name)) return null
-    return optBoolean(name)
 }
 
 internal const val LOCAL_REQUEST_ERROR_KEY = "_lyra_local_request_error"
